### 2023-09-10 #5823 @editorialbot
[PRE REVIEW]: Φ-ML: A Science-oriented Math and Neural Network Library for Jax, PyTorch, TensorFlow & NumPy
**Submitting author:** @holl- (<a href="http://orcid.org/0000-0001-9246-5195">Philipp Holl</a>)
**Repository:** https://github.com/tum-pbs/PhiML
**Branch with paper.md** (empty if default branch): joss-submission
**Version:** v1.0.0
