### 2023-09-11 #5823 @holl-
@arfon I rebased the `joss-submission` branch onto `main`. Had to force push to make it work, hope that's not a problem.
