### 2023-09-11 4240ca8
Add missing DOIs to paper.bib
