+++ paper.bib
@@ -53,7 +53,8 @@
   number={2},
   pages={180},
   year={2022},
-  publisher={MDPI}
+  publisher={MDPI},
+  doi = {10.3390/atmos13020180}
 }
 
 
@@ -75,7 +76,8 @@
   volume={52},
   pages={477--508},
   year={2020},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi = {10.1146/annurev-fluid-010719-060214},
 }
 
 
@@ -99,6 +101,7 @@
   number={2},
   pages={104},
   year={2020},
+  doi = {10.3847/1538-4357/abb9a7},
   publisher={IOP Publishing}
 }
 
@@ -110,6 +113,7 @@
   number={1},
   pages={87--93},
   year={2004},
+  doi = {10.1111/j.1365-2966.2004.07442.x},
   publisher={Blackwell Science Ltd Oxford, UK}
 }
 
@@ -121,6 +125,7 @@
   number={2},
   pages={50},
   year={2015},
+  doi = {10.1088/0004-637X/803/2/50},
   publisher={IOP Publishing}
 }
 
@@ -132,6 +137,7 @@
   volume={71},
   pages={804--818},
   year={2015},
+  doi = {10.1016/j.oregeorev.2015.01.001},
   publisher={Elsevier}
 }
 
@@ -165,6 +171,7 @@
   volume={378},
   pages={686--707},
   year={2019},
+  doi = {10.1016/j.jcp.2018.10.045},
   publisher={Elsevier}
 }
 
@@ -195,6 +202,7 @@
   booktitle={Pro Python},
   pages={301--302},
   year={2010},
+  doi = {10.1007/978-1-4302-2758-8_14},
   publisher={Springer}
 }
 
