### 2023-09-15 #5823 @gkthiruvathukal
@arfon Which command generates the cloc report? (the one that is triggered automatically on submissions)
