### 2023-09-20 593ffb5
Add argument already_padded to extrapolation functions
This could previously be used implicitly via **kwargs.
