+++ phiml/math/extrapolation.py
@@ -80,7 +80,7 @@ class Extrapolation:
         """
         raise NotImplementedError()
 
-    def pad(self, value: Tensor, widths: dict, **kwargs) -> Tensor:
+    def pad(self, value: Tensor, widths: dict, already_padded: Optional[dict] = None, **kwargs) -> Tensor:
         """
         Pads a tensor using values from `self.pad_values()`.
 
@@ -90,6 +90,9 @@ class Extrapolation:
         Args:
             value: `Tensor` to be padded
             widths: `dict` mapping `dim: str -> (lower: int, upper: int)`
+            already_padded: Used when padding a tensor with multiple extrapolations.
+                Contains all widths that have already been padded prior to this call.
+                This causes the shape of `value` to be different from the original tensor passed to `math.pad()`.
             kwargs: Additional keyword arguments for padding, passed on to `pad_values()`.
 
         Returns:
@@ -98,8 +101,8 @@ class Extrapolation:
         from ._trace import ShiftLinTracer
         if isinstance(value, ShiftLinTracer):
             lower = {dim: -lo for dim, (lo, _) in widths.items()}
-            return value.shift(lower, new_shape=value.shape.after_pad(widths), val_fun=lambda v: ZERO.pad(v, widths, **kwargs), bias_fun=lambda b: self.pad(b, widths, **kwargs))
-        already_padded = {}
+            return value.shift(lower, new_shape=value.shape.after_pad(widths), val_fun=lambda v: ZERO.pad(v, widths, already_padded=already_padded, **kwargs), bias_fun=lambda b: self.pad(b, widths, already_padded=already_padded, **kwargs))
+        already_padded = {} if already_padded is None else dict(already_padded)
         for dim, width in widths.items():
             assert (w > 0 for w in width), "Negative widths not allowed in Extrapolation.pad(). Use math.pad() instead."
             values = []
@@ -112,7 +115,7 @@ class Extrapolation:
             already_padded[dim] = width
         return value
 
-    def pad_values(self, value: Tensor, width: int, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def pad_values(self, value: Tensor, width: int, dim: str, upper_edge: bool, already_padded: Optional[dict] = None, **kwargs) -> Tensor:
         """
         Determines the values with which the given tensor would be padded at the specified using this extrapolation.
 
@@ -121,6 +124,9 @@ class Extrapolation:
             width: `int > 0`: Number of cells to pad along `dimension`.
             dim: Dimension name as `str`.
             upper_edge: `True` for upper edge, `False` for lower edge.
+            already_padded: Used when padding a tensor with multiple extrapolations.
+                Contains all widths that have already been padded prior to this call.
+                This causes the shape of `value` to be different from the original tensor passed to `math.pad()`.
 
         Returns:
             `Tensor` that can be concatenated to `value` along `dimension`
@@ -252,17 +258,8 @@ class ConstantExtrapolation(Extrapolation):
     def is_flexible(self) -> bool:
         return False
 
-    def pad(self, value: Tensor, widths: dict, **kwargs):
-        """
-        Pads a tensor using constant values.
-
-        Args:
-          value: `Tensor` to be padded
-          widths: name: str -> (lower: int, upper: int)}
-
-        Returns:
-            Padded `Tensor`
-        """
+    def pad(self, value: Tensor, widths: dict, already_padded: Optional[dict] = None, **kwargs) -> Tensor:
[…]
