### 2023-09-20 c1a0774
Fix constant extrapolation when already partially padded
