### 2023-09-22 e164710
Add invert to rotate_vector()
* Add unit test
