### 2023-09-22 #5823 @holl-
@gkthiruvathukal Seems to me that the new report covers the repository.

Also, I have added the MIT license to the top-level directory now so the above message `Failed to discover a valid open source license` should hopefully be fixed.
