### 2023-09-23 025a9d8
Minor: assert, annotation, formatting
