### 2023-09-28 #5823 @arfon
Also, please look at substantially reducing the length of your paper. As a guideline, JOSS papers should be typically around 1000 words. Thank you!
