### 2023-09-30 2d0be2e
Remove appendix sections from paper to reduce word count
