### 2023-10-03 08fcec5
Add extrapolation.get_normal/tangential()
