### 2023-10-03 831dc53
Add Shape.as_<type>()
