### 2023-10-04 90ccda9
Dimension.as_<type>() now makes no change if the dimension does not exist.
