### 2023-10-05 96e76cf
Update map()
* Add dims argument
* Add unwrap_scalars argument
* Move to _functional.py
* Fix stack() for strings
