+++ phiml/math/_magic_ops.py
@@ -160,7 +160,7 @@ def stack(values: Union[tuple, list, dict], dim: Shape, expand_values=False, **k
         else:
             values = [expand(v, all_dims.without(shape(v))) for v in values]
     else:
-        all_batch_dims = merge_shapes(*[batch(v) for v in values_], allow_varying_sizes=True)
+        all_batch_dims = merge_shapes(*[shape(v).batch for v in values_], allow_varying_sizes=True)
         if isinstance(values, dict):
             values = {k: expand(v, all_batch_dims.without(shape(v))) for k, v in values.items()}
         else:
@@ -277,7 +277,7 @@ def concat(values: Union[tuple, list], dim: Union[str, Shape], expand_values=Fal
             assert dim in shape(v), f"dim must be present in the shapes of all values bot got value {type(v).__name__} with shape {shape(v)}"
         for v in values[1:]:
             assert set(non_batch(v).names) == set(non_batch(values[0]).names), f"Concatenated values must have the same non-batch dimensions but got {non_batch(values[0])} and {non_batch(v)}"
-        all_batch_dims = merge_shapes(*[batch(v) for v in values])
+        all_batch_dims = merge_shapes(*[shape(v).batch for v in values])
         values = [expand(v, all_batch_dims) for v in values]
     # --- First try __concat__ ---
     for v in values:
