### 2023-10-06 eec9588
Remove private poisson_bracket()
