+++ phiml/math/_nd.py
@@ -777,49 +777,3 @@ def sample_subgrid(grid: Tensor, start: Tensor, size: Shape) -> Tensor:
             lower, upper = shift(grid, (0, 1), [dim], padding=None, stack_dim=None)
             grid = upper * upper_weight[i] + lower * lower_weight[i]
     return grid
-
-
-# Poisson Brackets
-
-
-def poisson_bracket(grid1, grid2):
-    if all([grid1.rank == grid2.rank == 2,
-            grid1.boundary == grid2.boundary == extrapolation.PERIODIC,
-            len(set(list(grid1.dx) + list(grid2.dx))) == 1]):
-        return _periodic_2d_arakawa_poisson_bracket(grid1.values, grid2.values, grid1.dx)
-    else:
-        raise NotImplementedError("\n".join([
-                                      "Not implemented for:"
-                                      f"ranks ({grid1.rank}, {grid2.rank}) != 2",
-                                      f"boundary ({grid1.boundary}, {grid2.boundary}) != {extrapolation.PERIODIC}",
-                                      f"dx uniform ({grid1.dx}, {grid2.dx})"
-                                  ]))
-
-
-def _periodic_2d_arakawa_poisson_bracket(tensor1: Tensor, tensor2: Tensor, dx: float):
-    """
-    Solves the poisson bracket using the Arakawa Scheme [tensor1, tensor2]
-    
-    Only works in 2D, with equal spaced grids, and periodic boundary conditions
-
-    Args:
-      tensor1(Tensor): first field in the poisson bracket
-      tensor2(Tensor): second field in the poisson bracket
-      dx(float): Grid size (equal in x-y)
-      tensor1: Tensor: 
-      tensor2: Tensor: 
-      dx: float: 
-
-    Returns:
-
-    """
-    zeta = math.pad(value=tensor1, widths={'x': (1, 1), 'y': (1, 1)}, mode=extrapolation.PERIODIC)
-    psi = math.pad(value=tensor2, widths={'x': (1, 1), 'y': (1, 1)}, mode=extrapolation.PERIODIC)
-    return (zeta.x[2:].y[1:-1] * (psi.x[1:-1].y[2:] - psi.x[1:-1].y[0:-2] + psi.x[2:].y[2:] - psi.x[2:].y[0:-2])
-            - zeta.x[0:-2].y[1:-1] * (psi.x[1:-1].y[2:] - psi.x[1:-1].y[0:-2] + psi.x[0:-2].y[2:] - psi.x[0:-2].y[0:-2])
-            - zeta.x[1:-1].y[2:] * (psi.x[2:].y[1:-1] - psi.x[0:-2].y[1:-1] + psi.x[2:].y[2:] - psi.x[0:-2].y[2:])
-            + zeta.x[1:-1].y[0:-2] * (psi.x[2:].y[1:-1] - psi.x[0:-2].y[1:-1] + psi.x[2:].y[0:-2] - psi.x[0:-2].y[0:-2])
-            + zeta.x[2:].y[0:-2] * (psi.x[2:].y[1:-1] - psi.x[1:-1].y[0:-2])
-            + zeta.x[2:].y[2:] * (psi.x[1:-1].y[2:] - psi.x[2:].y[1:-1])
-            - zeta.x[0:-2].y[2:] * (psi.x[1:-1].y[2:] - psi.x[0:-2].y[1:-1])
-            - zeta.x[0:-2].y[0:-2] * (psi.x[0:-2].y[1:-1] - psi.x[1:-1].y[0:-2])) / (12 * dx ** 2)
