### 2023-10-06 9b2f8eb
Add epsilon to vec_normalize()
