### 2023-10-06 0e7f55a
Rotation matrix from axis-angle
