+++ phiml/math/_nd.py
@@ -178,44 +178,57 @@ def rotate_vector(vector: math.Tensor, angle: Optional[Union[float, math.Tensor]
     assert 'vector' in vector.shape, f"vector must have exactly a channel dimension named 'vector'"
     if angle is None:
         return vector
-    matrix = rotation_matrix(angle, matrix_dim_2d=channel(vector))
+    matrix = rotation_matrix(angle, matrix_dim=channel(vector))
     if invert:
         matrix = rename_dims(matrix, '~vector,vector', math.concat_shapes(channel('vector'), dual('vector')))
     assert matrix.vector.dual.size == vector.vector.size, f"Rotation matrix from {angle.shape} is {matrix.vector.dual.size}D but vector {vector.shape} is {vector.vector.size}D."
     return matrix @ vector
 
 
-def rotation_matrix(angle: Union[float, math.Tensor], matrix_dim_2d=channel(vector='x,y')):
+def rotation_matrix(x: Union[float, math.Tensor], matrix_dim=channel('vector')):
     """
     Create a 2D or 3D rotation matrix from the corresponding angle(s).
 
     Args:
-        angle: Either scalar angle for a 2D rotation or euler angles along `vector` in 3D.
+        x:
+            2D: scalar angle
+            3D: Either vector pointing along the rotation axis with rotation angle as length or Euler angles.
+            Euler angles need to be laid out along a `angle` channel dimension with dimension names listing the spatial dimensions.
+            E.g. a 90° rotation about the z-axis is represented by `vec('angles', x=0, y=0, z=PI/2)`.
             If a rotation matrix is passed for `angle`, it is returned without modification.
-        matrix_dim_2d: Matrix dimension for 2D rotations. In 3D, the channel dimension of angle is used.
+        matrix_dim: Matrix dimension for 2D rotations. In 3D, the channel dimension of angle is used.
 
     Returns:
-        Matrix with one channel and one dual dimension as well as all non-channel dimensions of `angle`.
-        In 2D, the channel and dual dimensions are derived from `matrix_dim_2d` and in 3D, the channel dimension of `angle` is used instead.
+        Matrix containing `matrix_dim` in primal and dual form as well as all non-channel dimensions of `x`.
     """
-    if isinstance(angle, Tensor) and '~vector' in angle.shape and 'vector' in angle.shape.channel and angle.shape.get_size('~vector') == angle.shape.get_size('vector'):
-        return angle  # already a rotation matrix
-    if 'vector' not in shape(angle) or shape(angle).get_size('vector') == 1:  # 1D rotation
-        sin = wrap(math.sin(angle))
-        cos = wrap(math.cos(angle))
-        return wrap([[cos, -sin], [sin, cos]], matrix_dim_2d, matrix_dim_2d.as_dual())
-    else:
-        assert channel(angle).rank == 1 and channel(angle).size == 3, f"angle for 3D rotations needs to be a 3-vector but got {angle}"
-        s1, s2, s3 = math.sin(angle).vector  # phi, theta, psi
-        c1, c2, c3 = math.cos(angle).vector
-        return wrap([[c1 * c2, c1 * s2 * s3 - s1 * c3, c1 * s2 * c3 + s1 * s3],
-                     [s1 * c2, s1 * s2 * s3 + c1 * c3, s1 * s2 * c3 - c1 * s3],
[…]
