### 2023-10-08 cec41e7
Legacy compatibility for mixed extrapolation serialization
