### 2023-10-08 9e22291
Fix stack() for empty dims
