### 2023-10-09 25e59fd
Add Shape.assert_all_sizes_defined()
