### 2023-10-09 bae47dc
Improvements and fixes for map(), broadcast()
* Fix map() default dims
* Keep positional arguments in map()
* Support DimFilter in map()
* Fix map_ import
* Fix overhead in linear solves
