### 2023-10-10 416a9b5
Add angle(), rename degrees()
