+++ phiml/math/_ops.py
@@ -2,7 +2,7 @@ import functools
 import math
 import warnings
 from numbers import Number
-from typing import Tuple, Callable, Any, Union, Optional, Dict, Collection
+from typing import Tuple, Callable, Any, Union, Optional, Dict, Collection, TypeVar
 
 import numpy as np
 
@@ -14,7 +14,7 @@ from ._shape import (Shape, EMPTY_SHAPE,
 from ._sparse import CompressedSparseMatrix, dot_compressed_dense, dense, SparseCoordinateTensor, dot_coordinate_dense, get_format, to_format, stored_indices, tensor_like, sparse_dims, same_sparsity_pattern, is_sparse, sparse_dot
 from ._tensors import (Tensor, wrap, tensor, broadcastable_native_tensors, NativeTensor, TensorStack,
                        custom_op2, compatible_tensor, variable_attributes, disassemble_tree, assemble_tree,
-                       is_scalar, Layout, expand_tensor)
+                       is_scalar, Layout, expand_tensor, TensorOrTree)
 from ..backend import default_backend, choose_backend, Backend, get_precision, convert as b_convert, BACKENDS, NoBackendFound, ComputeDevice, NUMPY
 from ..backend._dtype import DType, combine_types
 from .magic import PhiTreeNode, Shapable
@@ -1754,7 +1754,7 @@ def _backend_op1(x, unbound_method) -> Union[Tensor, PhiTreeNode]:
         return y
 
 
-def abs_(x) -> Union[Tensor, PhiTreeNode]:
+def abs_(x: TensorOrTree) -> TensorOrTree:
     """
     Computes *||x||<sub>1</sub>*.
     Complex `x` result in matching precision float values.
@@ -1771,7 +1771,7 @@ def abs_(x) -> Union[Tensor, PhiTreeNode]:
     return _backend_op1(x, Backend.abs)
 
 
-def sign(x) -> Union[Tensor, PhiTreeNode]:
+def sign(x: TensorOrTree) -> TensorOrTree:
     """
     The sign of positive numbers is 1 and -1 for negative numbers.
     The sign of 0 is undefined.
@@ -1785,37 +1785,37 @@ def sign(x) -> Union[Tensor, PhiTreeNode]:
     return _backend_op1(x, Backend.sign)
 
 
-def round_(x) -> Union[Tensor, PhiTreeNode]:
+def round_(x: TensorOrTree) -> TensorOrTree:
     """ Rounds the `Tensor` or `phiml.math.magic.PhiTreeNode` `x` to the closest integer. """
     return _backend_op1(x, Backend.round)
 
 
-def ceil(x) -> Union[Tensor, PhiTreeNode]:
+def ceil(x: TensorOrTree) -> TensorOrTree:
     """ Computes *⌈x⌉* of the `Tensor` or `phiml.math.magic.PhiTreeNode` `x`. """
     return _backend_op1(x, Backend.ceil)
 
 
-def floor(x) -> Union[Tensor, PhiTreeNode]:
+def floor(x: TensorOrTree) -> TensorOrTree:
     """ Computes *⌊x⌋* of the `Tensor` or `phiml.math.magic.PhiTreeNode` `x`. """
     return _backend_op1(x, Backend.floor)
 
 
-def sqrt(x) -> Union[Tensor, PhiTreeNode]:
+def sqrt(x: TensorOrTree) -> TensorOrTree:
     """ Computes *sqrt(x)* of the `Tensor` or `phiml.math.magic.PhiTreeNode` `x`. """
     return _backend_op1(x, Backend.sqrt)
 
[…]
