+++ phiml/math/__init__.py
@@ -56,8 +56,8 @@ from ._ops import (
     maximum, minimum, clip,
     sqrt, exp, log, log2, log10, sigmoid, soft_plus,
     sin, cos, tan, sinh, cosh, tanh, arcsin, arccos, arctan, arcsinh, arccosh, arctanh, log_gamma, factorial,
-    to_float, to_int32, to_int64, to_complex, imag, real, conjugate,
-    degrees,
+    to_float, to_int32, to_int64, to_complex, imag, real, conjugate, angle,
+    radians_to_degrees, degrees_to_radians,
     boolean_mask,
     is_finite, is_nan, is_inf,
     closest_grid_values, grid_sample, scatter, gather,
