### 2023-10-10 c4a2086
Add conditional extrapolation
* Implement missing Extrapolation methods
* extrapolation.where()
* math.where()
* Fix extrapolation.map()
* Fix __repr__, __eq__
* Fix __ne__ for all Extrapolations
