+++ phiml/math/extrapolation.py
@@ -6,6 +6,7 @@ Extrapolations are an important part of sampled fields such as grids.
 See the documentation at https://tum-pbs.github.io/PhiML/Fields.html#extrapolations .
 """
 import warnings
+from abc import ABC
 from numbers import Number
 from typing import Union, Dict, Callable, Tuple, Optional, Hashable, Sequence
 
@@ -13,7 +14,7 @@ from ..backend._backend import get_spatial_derivative_order
 from ..backend import choose_backend
 from ._shape import Shape, channel, spatial, EMPTY_SHAPE, merge_shapes, dual, non_dual, instance
 from ._magic_ops import concat, stack, expand, rename_dims
-from ._tensors import Tensor, NativeTensor, TensorStack, wrap
+from ._tensors import Tensor, NativeTensor, TensorStack, wrap, to_dict as tensor_to_dict, from_dict as tensor_from_dict
 from . import _ops as math  # TODO this executes _ops.py, can we avoid this?
 
 
@@ -185,6 +186,12 @@ class Extrapolation:
 
     def _getitem_with_domain(self, item: dict, dim: str, upper_edge: bool, all_dims: Sequence[str]):
         return self[item]
+
+    def __eq__(self, other):
+        raise NotImplementedError(self.__class__)
+
+    def __ne__(self, other):
+        return not self == other
 
     def __abs__(self):
         raise NotImplementedError(self.__class__)
@@ -398,7 +405,7 @@ class ConstantExtrapolation(Extrapolation):
         return ConstantExtrapolation(-self.value)
 
 
-class _CopyExtrapolation(Extrapolation):
+class _CopyExtrapolation(Extrapolation, ABC):
 
     @property
     def shape(self):
@@ -617,6 +624,7 @@ class _ZeroGradient(_CopyExtrapolation):
 
 class _PeriodicExtrapolation(_CopyExtrapolation):
     """ Periodic extrapolation in n dimensions. """
+
     def __repr__(self):
         return 'periodic'
 
@@ -652,6 +660,9 @@ class _PeriodicExtrapolation(_CopyExtrapolation):
     def shortest_distance(self, start: Tensor, end: Tensor, domain_size: Tensor):
         dx = end - start
         return (dx + domain_size / 2) % domain_size - domain_size / 2
+
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+        raise AssertionError("Periodic extrapolation cannot be used with sparse connectivity. Instead, add periodicity to the connectivity matrix.")
 
 
 class _SymmetricExtrapolation(_CopyExtrapolation):
@@ -708,30 +719,19 @@ class _SymmetricExtrapolation(_CopyExtrapolation):
         return result
 
     def _lower_mask(self, shape, widths, bound_dim, bound_lo, bound_hi, i):
-        # key = (shape, tuple(widths.keys()), tuple(widths.values()), bound_dim, bound_lo, bound_hi, i)
-        # if key in _BoundaryExtrapolation._CACHED_LOWER_MASKS:
-        #     result = math.tensor(_BoundaryExtrapolation._CACHED_LOWER_MASKS[key])
-        #     _BoundaryExtrapolation._CACHED_LOWER_MASKS[key] = result
-        #     return result
-        # else:
-            mask = ZERO.pad(math.zeros(shape), {bound_dim: (bound_lo - i - 1, 0)})
-            mask = ONE.pad(mask, {bound_dim: (1, 0)})
-            mask = ZERO.pad(mask, {dim: (i, bound_hi) if dim == bound_dim else (lo, hi) for dim, (lo, hi) in widths.items()})
-            # _BoundaryExtrapolation._CACHED_LOWER_MASKS[key] = mask
[…]
