### 2023-10-12 9263f95
Support basic rename_dims() inside traced linear functions
