+++ phiml/math/_trace.py
@@ -7,7 +7,7 @@ import numpy as np
 from ..backend import choose_backend, NUMPY, Backend
 from ._ops import choose_backend_t
 from ._shape import Shape, parse_dim_order, merge_shapes, spatial, instance, batch, concat_shapes, EMPTY_SHAPE, dual, channel, non_batch
-from ._magic_ops import stack, expand
+from ._magic_ops import stack, expand, rename_dims
 from ._tensors import Tensor, wrap, disassemble_tree, disassemble_tensors, assemble_tree, TensorStack, may_vary_along
 from ._sparse import SparseCoordinateTensor
 from . import _ops as math
@@ -24,7 +24,7 @@ class ShiftLinTracer(Tensor):
     Dimensions not contained in any `val` Tensor are treated as independent (batch dimensions).
     """
 
-    def __init__(self, source: TracerSource, values_by_shift: dict, shape: Shape, bias: Tensor):
+    def __init__(self, source: TracerSource, values_by_shift: dict, shape: Shape, bias: Tensor, renamed: Dict[str, str]):
         """
         Args:
             source: placeholder tensor
@@ -37,12 +37,14 @@ class ShiftLinTracer(Tensor):
                 However, the bias can be subtracted from the solution vector when solving a linear system, allowing this function to be solved with regular linear system solvers.
         """
         assert isinstance(source, TracerSource)
+        assert isinstance(renamed, dict)
         self.source = source
         self.val: Dict[Shape, Tensor] = simplify_add(values_by_shift)
         for shift_ in self.val.keys():
             assert shift_.only(sorted(shift_.names), reorder=True) == shift_
         self.bias = bias
         self._shape = shape
+        self._renamed = renamed  # new_name -> old_name
 
     def __repr__(self):
         return f"Linear tracer {self._shape}"
@@ -77,8 +79,9 @@ class ShiftLinTracer(Tensor):
     def shape(self):
         return self._shape
 
-    def _with_shape_replaced(self, new_shape):
-        raise NotImplementedError()
+    def _with_shape_replaced(self, new_shape: Shape):
+        renamed = {new_dim: self._renamed[old_dim] for old_dim, new_dim in zip(self._shape.names, new_shape.names)}
+        return ShiftLinTracer(self.source, self.val, new_shape, self.bias, renamed)
 
     @property
     def _is_tracer(self) -> bool:
@@ -116,13 +119,13 @@ class ShiftLinTracer(Tensor):
                     shift = shift._replace_single_size(dim, shift.get_size(dim) + delta) if dim in shift else shift._expand(spatial(**{dim: delta}))
             val[shift.only(sorted(shift.names), reorder=True)] = val_fun(values)
         bias = bias_fun(self.bias)
-        return ShiftLinTracer(self.source, val, new_shape, bias)
+        return ShiftLinTracer(self.source, val, new_shape, bias, self._renamed)
 
     def _unstack(self, dimension):
         raise NotImplementedError()
 
     def __neg__(self):
-        return ShiftLinTracer(self.source, {shift: -values for shift, values in self.val.items()}, self._shape, -self.bias)
+        return ShiftLinTracer(self.source, {shift: -values for shift, values in self.val.items()}, self._shape, -self.bias, self._renamed)
 
     def _op1(self, native_function):
         # __neg__ is the only proper linear op1 and is implemented above.
@@ -168,7 +171,7 @@ class ShiftLinTracer(Tensor):
                     else:
                         values[dim_shift] = other_values
             bias = operator(self.bias, other.bias)
-            return ShiftLinTracer(self.source, values, self._shape, bias)
[…]
