### 2023-10-14 ef7e03a
Add Shape.resolve_index()
