+++ phiml/math/_ops.py
@@ -2,7 +2,7 @@ import functools
 import math
 import warnings
 from numbers import Number
-from typing import Tuple, Callable, Any, Union, Optional, Dict, Collection, TypeVar
+from typing import Tuple, Callable, Any, Union, Optional, Dict, Collection, TypeVar, List, Sequence
 
 import numpy as np
 
@@ -17,7 +17,7 @@ from ._tensors import (Tensor, wrap, tensor, broadcastable_native_tensors, Nativ
                        is_scalar, Layout, expand_tensor, TensorOrTree)
 from ..backend import default_backend, choose_backend, Backend, get_precision, convert as b_convert, BACKENDS, NoBackendFound, ComputeDevice, NUMPY
 from ..backend._dtype import DType, combine_types
-from .magic import PhiTreeNode, Shapable
+from .magic import PhiTreeNode, Shapable, Shaped
 
 
 def choose_backend_t(*values, prefer_default=False) -> Backend:
@@ -883,15 +883,19 @@ def concat_tensor(values: Union[tuple, list], dim: str) -> Tensor:
     return result
 
 
-def pad(value: Tensor, widths: Union[dict, tuple], mode: Union['e_.Extrapolation', Tensor, Number, str, dict] = 0, **kwargs) -> Tensor:
+def pad(value: Tensor, widths: Union[dict, tuple, list], mode: Union['e_.Extrapolation', Tensor, Number, str, dict] = 0, **kwargs) -> Tensor:
     """
     Pads a tensor along the specified dimensions, determining the added values using the given extrapolation.
     Unlike `Extrapolation.pad()`, this function can handle negative widths which slice off outer values.
 
     Args:
         value: `Tensor` to be padded
-        widths: `dict` mapping dimension name (`str`) to `(lower, upper)`
-            where `lower` and `upper` are `int` that can be positive (pad), negative (slice) or zero (pass).
+        widths: Number of values to add at the edge of `value`. Negative values can be used to slice off edge values. Must be one of the following:
+
+            * `tuple` containing `(lower: int, upper: int)`. This will pad all non-batch dimensions by `lower` and `upper` at the lower and upper edge, respectively.
+            * `dict` mapping `dim: str -> (lower: int, upper: int)`
+            * Sequence of slicing `dict`s. This will add all values specified by the slicing dicts and is the inverse operation to `slice_off`. Exactly one value in each slicing dict must be a `slice` object.
+
         mode: Padding mode used to determine values added from positive `widths`.
             Must be one of the following: `Extrapolation`, `Tensor` or number for constant extrapolation, name of extrapolation as `str`.
         kwargs: Additional padding arguments.
@@ -907,7 +911,10 @@ def pad(value: Tensor, widths: Union[dict, tuple], mode: Union['e_.Extrapolation
         >>> math.pad(math.ones(spatial(x=10, y=10)), {'x': (1, -1)}, 0)
         (xˢ=10, yˢ=10) 0.900 ± 0.300 (0e+00...1e+00)
     """
+    mode = e_.as_extrapolation(mode)
     if isinstance(widths, (tuple, list)):
+        if len(widths) == 0 or isinstance(widths[0], dict):  # add sliced-off slices
+            return _pad_slices(value, widths, mode, **kwargs)
         if len(widths) == 2 and isinstance(widths[0], int) and isinstance(widths[1], int):  # (lower, upper)
             assert non_batch(value).rank == 1, f"Can only pad 1D tensors (excluding batch dims) when widths=(lower, upper) but got {shape(value)} and widths={widths}"
[…]
