### 2023-10-14 799a177
Fix type hints for laplace, spatial_gradient
