### 2023-10-19 322dca6
Fix get_backend() for profiling
* Add unit test
