### 2023-10-19 6b8e7df
Optimize close() for non-uniform tensors
