### 2023-10-19 b85e9cb
Add two_sided to as_extrapolation()
* Refactor sparse_pad_values
