+++ phiml/math/extrapolation.py
@@ -137,7 +137,7 @@ class Extrapolation:
         """
         raise NotImplementedError(self.__class__)
 
-    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, **kwargs) -> Tensor:
         raise NotImplementedError(self.__class__)
 
     def transform_coordinates(self, coordinates: Tensor, shape: Shape, **kwargs) -> Tensor:
@@ -313,7 +313,7 @@ class ConstantExtrapolation(Extrapolation):
         else:
             return math.wrap(0)
 
-    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, **kwargs) -> Tensor:
         return math.expand(self.value, dual(connectivity) & non_dual(value))
 
     def __eq__(self, other):
@@ -608,13 +608,13 @@ class _ZeroGradient(_CopyExtrapolation):
             # _BoundaryExtrapolation._CACHED_UPPER_MASKS[key] = mask
             return mask
 
-    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, **kwargs) -> Tensor:
         from ._sparse import stored_indices
         from ._ops import arange
         dual_dim = dual(value).name
-        primal_dim = dual(value).name[1:]
         # --- Gather the edge values ---
         indices = stored_indices(connectivity, invalid='discard')
+        primal_dim = [n for n in indices.index.item_names if not n.startswith('~')][0]
         gathered = value[{dual_dim: indices[primal_dim]}]
         # --- Scatter, but knowing there is only one entry per row & col, we can simply permute ---
         inv_perm = arange(dual(connectivity))[{dual_dim: indices[dual_dim]}]
@@ -661,7 +661,7 @@ class _PeriodicExtrapolation(_CopyExtrapolation):
         dx = end - start
         return (dx + domain_size / 2) % domain_size - domain_size / 2
 
-    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, **kwargs) -> Tensor:
         raise AssertionError("Periodic extrapolation cannot be used with sparse connectivity. Instead, add periodicity to the connectivity matrix.")
 
 
@@ -730,7 +730,7 @@ class _SymmetricExtrapolation(_CopyExtrapolation):
         mask = ZERO.pad(mask, {dim: (bound_lo, i) if dim == bound_dim else (lo, hi) for dim, (lo, hi) in widths.items()})
         return mask
 
-    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, upper_edge: bool, **kwargs) -> Tensor:
+    def sparse_pad_values(self, value: Tensor, connectivity: Tensor, dim: str, **kwargs) -> Tensor:
         raise NotImplementedError
 
 
@@ -773,7 +773,7 @@ class _AntiSymmetricExtrapolation(_SymmetricExtrapolation):
                 result -= boundary  # this does basically nothing if value is the identity
         return result
 
[…]
