### 2023-10-20 ab2805a
Sparse pad shape check
