### 2023-10-23 0157764
Add alias math.as_extrapolation()
