### 2023-10-23 56223ca
Fix debug detection in Tensor formatting
