+++ phiml/math/_tensors.py
@@ -2268,7 +2268,7 @@ def check_is_printing():
     import traceback, sys
     stack = traceback.extract_stack()
     for frame in stack:
-        if "_pydevd_bundle\\pydevd_xml.py" in frame.filename:
+        if "_pydevd_bundle\\pydevd_xml.py" in frame.filename or "_pydevd_bundle/pydevd_xml.py" in frame.filename:
             return False
     for frame in stack:
         if frame.line.strip().startswith('print('):
