### 2023-10-23 5586da9
boolean_mask() for non-Tensor arguments and sparse matrices
