### 2023-10-23 e7ad944
Fix assert_close() for sparse tensors
