### 2023-10-23 b2d2115
Sparse dot product for non-sparse dims only
