### 2023-10-23 75a52dd
Use partial matmul when only some dims match
