### 2023-10-27 6e8fb5e
Add when_available()
