+++ phiml/math/__init__.py
@@ -91,6 +91,7 @@ from ._functional import (
     identity,
     trace_check,
     map_ as map,
+    when_available,
 )
 
 from ._optimize import solve_linear, solve_nonlinear, minimize, Solve, SolveInfo, ConvergenceException, NotConverged, Diverged, SolveTape, factor_ilu
