### 2023-10-27 61b9a4f
Fix TensorFlow numpy_call
