### 2023-10-27 bd88d90
Add perf_counter()
