### 2023-11-07 1a70230
Allow unstack() with empty dims
