### 2023-11-08 e81c0d5
Trace sparse matrices in linear functions
* Clean up tracing state on error
* Fix linear trace with renamed dims
