### 2023-11-10 e248a0f
Disable unit tests on Python < 3.8 as Jax no longer supports it
