+++ .github/workflows/unit-tests.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ['3.7', '3.8', '3.11']
+        python-version: ['3.8', '3.11']
 
     steps:
       - name: Checkout 🛎️
