### 2023-11-11 398d0c7
Fix @jit_compile_linear returning Tensors instead of composite objects
