### 2023-11-11 4cd9a16
Implement sparse COO features
* compression with duplicate entries
* COO-COO addition / subtraction
