### 2023-11-11 69c4171
Linear tracing improvements
* Fix too-strict tracing assertion
