+++ phiml/math/_trace.py
@@ -6,7 +6,7 @@ import numpy as np
 
 from ..backend import choose_backend, NUMPY, Backend
 from ._ops import choose_backend_t, concat_tensor, scatter, zeros_like
-from ._shape import Shape, parse_dim_order, merge_shapes, spatial, instance, batch, concat_shapes, EMPTY_SHAPE, dual, channel, non_batch, primal
+from ._shape import Shape, parse_dim_order, merge_shapes, spatial, instance, batch, concat_shapes, EMPTY_SHAPE, dual, channel, non_batch, primal, non_channel, DEBUG_CHECKS
 from ._magic_ops import stack, expand, rename_dims, unpack_dim
 from ._tensors import Tensor, wrap, disassemble_tree, disassemble_tensors, assemble_tree, TensorStack, may_vary_along, discard_constant_dims, variable_shape
 from ._sparse import SparseCoordinateTensor, is_sparse, sparse_dims, same_sparsity_pattern, sparse_tensor
@@ -40,8 +40,9 @@ class ShiftLinTracer(Tensor):
         assert isinstance(renamed, dict)
         self._source = source
         self.val: Dict[Shape, Tensor] = simplify_add(values_by_shift)
-        for shift_ in self.val.keys():
+        for shift_, v in self.val.items():
             assert shift_.only(sorted(shift_.names), reorder=True) == shift_
+            assert v.shape.only(shape, reorder=True) == shape.only(v.shape), f"Tracer with shape {shape} must have matching values but got {v.shape}"  # values must match shape
         self._bias = bias
         self._shape = shape
         self._renamed = renamed  # new_name -> old_name
@@ -75,7 +76,7 @@ class ShiftLinTracer(Tensor):
         new_shape = math.zeros(self._shape)[selection].shape
         return self.shift(starts, new_shape, lambda v: v[selection], lambda b: b[selection])
 
-    def shift(self, shifts: dict,
+    def shift(self, shifts: Dict[str, int],
               new_shape: Shape,
               val_fun: Callable,
               bias_fun: Callable = None):
@@ -124,10 +125,9 @@ class ShiftLinTracer(Tensor):
              operator: Callable,
              native_function: Callable,
              op_name: str = 'unknown',
-             op_symbol: str = '?') -> 'ShiftLinTracer':
+             op_symbol: str = '?') -> Tensor:
         if isinstance(other, SparseLinTracer):
-            sparse_self = self.to_sparse_tracer()
-            return sparse_self._op2(other, operator, native_function, op_name, op_symbol)
+            return to_sparse_tracer(self, other)._op2(other, operator, native_function, op_name, op_symbol)
         assert op_symbol in '+-*/', f"Unsupported operation encountered while tracing linear function: {native_function}"
         zeros_for_missing_self = op_name not in ['add', 'radd', 'rsub']  # perform `operator` where `self == 0`
         zeros_for_missing_other = op_name not in ['add', 'radd', 'sub']  # perform `operator` where `other == 0`
@@ -208,18 +208,21 @@ class GatherLinTracer(Tensor):
         assert bias.shape in shape
         assert selection is None or selection.dtype.kind == int
         assert diag.shape in shape
+        assert selection is None or selection.shape.volume > 0
         self._source = source
         self._diag = diag  # full matrix or diagonal elements only
         self._bias = bias  # matches self.shape
         self._selection = selection  # Can index one or multiple dimensions of the source. Must retain source dimensions.
         self._shape = shape
         self._renamed = renamed  # dims renamed before matrix mul. new_name -> old_name
+        if DEBUG_CHECKS:
+            if selection is not None:
[…]
