### 2023-11-13 48aff84
Fix default mul_coo_dense()
