### 2023-11-13 684cd1a
Fix stack() for highly non-uniform tensors
