### 2023-11-13 1c25305
Improved sparse tracing, fixes
* Trace dot products
