+++ phiml/math/_ops.py
@@ -11,7 +11,7 @@ from ._magic_ops import expand, pack_dims, unpack_dim, cast, copy_with, value_at
 from ._shape import (Shape, EMPTY_SHAPE,
                      spatial, batch, channel, instance, merge_shapes, parse_dim_order, concat_shapes,
                      IncompatibleShapes, DimFilter, non_batch, dual, non_channel, shape)
-from ._sparse import CompressedSparseMatrix, dot_compressed_dense, dense, SparseCoordinateTensor, dot_coordinate_dense, get_format, to_format, stored_indices, tensor_like, sparse_dims, same_sparsity_pattern, is_sparse, sparse_dot
+from ._sparse import CompressedSparseMatrix, dot_compressed_dense, dense, SparseCoordinateTensor, dot_coordinate_dense, get_format, to_format, stored_indices, tensor_like, sparse_dims, same_sparsity_pattern, is_sparse, sparse_dot, sparse_sum, sparse_gather
 from ._tensors import (Tensor, wrap, tensor, broadcastable_native_tensors, NativeTensor, TensorStack,
                        custom_op2, compatible_tensor, variable_attributes, disassemble_tree, assemble_tree,
                        is_scalar, Layout, expand_tensor, TensorOrTree, discard_constant_dims)
@@ -1314,36 +1314,7 @@ def _sum(value: Tensor, dims: Shape) -> Tensor:
         reduced_inners = [_sum(t, dims.without(value._stack_dim)) for t in value._tensors]
         return functools.reduce(lambda x, y: x + y, reduced_inners) if value._stack_dim in dims else TensorStack(reduced_inners, value._stack_dim)
     elif isinstance(value, (CompressedSparseMatrix, SparseCoordinateTensor)):
-        if value.sparse_dims in dims:  # reduce all sparse dims
-            return _sum(value._values, dims.without(value.sparse_dims) & instance(value._values))
-        value_only_dims = dims.only(value._values.shape).without(value.sparsity_batch)
-        if value_only_dims:
-            value = value._with_values(_sum(value._values, value_only_dims))
-        dims = dims.without(value_only_dims)
-        if not dims:
-            return value
-        if isinstance(value, CompressedSparseMatrix):
-            if value._compressed_dims in dims and value._uncompressed_dims.isdisjoint(dims):  # We can ignore the pointers
-                result_base = value.shape.without(value._compressed_dims)
-                return scatter(result_base, value._indices, value._values, mode='add', outside_handling='undefined')
-            elif value.sparse_dims.only(dims):  # reduce some sparse dims
-                return dot(value, dims, ones(dims), dims)  # this is what SciPy does in both axes, actually.
-            return value
-            # first sum value dims that are not part of indices
-        else:
-            assert isinstance(value, SparseCoordinateTensor)
-            if value._dense_shape in dims:  # sum all sparse dims
-                v_dims = dims.without(value._dense_shape) & instance(value._values)
-                return _sum(value._values, v_dims)
-            else:
-                result_base = zeros(value.shape.without(dims))
-                remaining_sparse_dims = value._dense_shape.without(dims)
-                indices = value._indices.vector[remaining_sparse_dims.names]
-                if remaining_sparse_dims.rank == 1:  # return dense result
-                    result = scatter(result_base, indices, value._values, mode='add', outside_handling='undefined')
-                    return result
-                else:  # return sparse result
-                    raise NotImplementedError
+        return sparse_sum(value, dims)
     else:
         raise ValueError(type(value))
 
@@ -1782,13 +1753,9 @@ def dot(x: Tensor,
     x_dims = x.shape.only(x_dims)
[…]
