+++ phiml/math/_trace.py
@@ -3,13 +3,14 @@ from typing import Callable, Dict, Set, Tuple, Union, Any, Optional, Sequence, L
 
 import numpy
 import numpy as np
+from scipy.sparse import csr_matrix
 
 from ..backend import choose_backend, NUMPY, Backend
 from ._ops import choose_backend_t, concat_tensor, scatter, zeros_like
 from ._shape import Shape, parse_dim_order, merge_shapes, spatial, instance, batch, concat_shapes, EMPTY_SHAPE, dual, channel, non_batch, primal, non_channel, DEBUG_CHECKS
-from ._magic_ops import stack, expand, rename_dims, unpack_dim
+from ._magic_ops import stack, expand, rename_dims, unpack_dim, unstack
 from ._tensors import Tensor, wrap, disassemble_tree, disassemble_tensors, assemble_tree, TensorStack, may_vary_along, discard_constant_dims, variable_shape
-from ._sparse import SparseCoordinateTensor, is_sparse, sparse_dims, same_sparsity_pattern, sparse_tensor
+from ._sparse import SparseCoordinateTensor, is_sparse, sparse_dims, same_sparsity_pattern, sparse_tensor, stored_indices, stored_values, add_sparse_batch_dim
 from . import _ops as math
 from ..backend._dtype import combine_types
 
@@ -126,6 +127,8 @@ class ShiftLinTracer(Tensor):
              native_function: Callable,
              op_name: str = 'unknown',
              op_symbol: str = '?') -> Tensor:
+        if is_sparse(other):
+            return NotImplemented
         if isinstance(other, SparseLinTracer):
             return to_sparse_tracer(self, other)._op2(other, operator, native_function, op_name, op_symbol)
         assert op_symbol in '+-*/', f"Unsupported operation encountered while tracing linear function: {native_function}"
@@ -175,9 +178,9 @@ class ShiftLinTracer(Tensor):
     def _spec_dict(self) -> dict:
         raise LinearTraceInProgress(self)
 
-    def _matmul(self, matrix: Tensor, sdims: Shape, ddims: Shape):
+    def _matmul(self, self_dims: Shape, matrix: Tensor, matrix_dims: Shape) -> Tensor:
         if is_sparse(matrix):
-            return to_gather_tracer(self).matmul(matrix, sdims, ddims)
+            return to_gather_tracer(self).matmul(self_dims, matrix, matrix_dims)
         raise NotImplementedError
 
     def _gather(self, indices: Tensor) -> Tensor:
@@ -224,12 +227,14 @@ class GatherLinTracer(Tensor):
     def __repr__(self):
         return f"{self.__class__.__name__} {self._shape}"
 
-    def _matmul(self, matrix: Tensor, mdims: Shape, ddims: Shape):
-        shape = matrix.shape.without(mdims) & self._shape.without(ddims)
-        matrix *= self._matrix
-        matrix = rename_dims(matrix, mdims, rename_dims(ddims, [*self._renamed.keys()], [*self._renamed.values()]).as_dual())
-        renamed = {n: o for n, o in self._renamed.items() if n not in ddims}
-        return SparseLinTracer(self._source, matrix, self._bias, shape, self._selection, renamed)
+    def _matmul(self, self_dims: Shape, matrix: Tensor, matrix_dims: Shape) -> Tensor:
+        shape = matrix.shape.without(matrix_dims) & self._shape.without(self_dims)
+        if self_dims not in self._diag.shape:  # self is constant along self_dims
+            matrix = math.sum_(matrix, matrix_dims)
+        diag = matrix * self._diag
[…]
