### 2023-11-15 0ea6084
Fix item names when retyping using rename_dims
