### 2023-11-20 e4af47b
Add neighbor_reduce() and partial functions
* Add unit test
