### 2023-11-20 6c366e1
Allow scalar int tensors in meshgrid()
