### 2023-11-20 4bb4a04
Support other types for extrap on grid_sample()
