### 2023-11-20 a9bb566
Add at_min_neighbor(), at_max_neighbor()
