### 2023-11-21 4a2ae8b
Add index_shift()
* Update shift() docstring
* Add index_shift_widths()
* Add unit tests
