### 2023-11-23 417734b
Support trees in gather()
