### 2023-11-23 c782bf9
Add allow_infinite to vec_normalize()
