### 2023-11-23 b120e4d
Fix at_max for values with different dims than key
