### 2023-11-24 f0f4086
Fix dtype for bool
