+++ phiml/backend/jax/_jax_backend.py
@@ -564,6 +564,8 @@ class JaxBackend(Backend):
             return jnp.fft.ifftn(k, axes=axes).astype(k.dtype)
 
     def dtype(self, array) -> DType:
+        if isinstance(array, bool):
+            return DType(bool)
         if isinstance(array, int):
             return DType(int, 32)
         if isinstance(array, float):
