### 2023-11-24 1bf1dea
Fix invert (~) operator for bool
