### 2023-11-24 a5014f8
Fix Sliceable instance check
