### 2023-11-24 8948224
Support objects without __getitem__ in map()
