### 2023-11-24 cd85f07
Fix at_max, at_min for multi-dim reduce
