### 2023-11-24 cd89b2b
Add Tensor op1 return type annotations
