### 2023-11-24 8dfe33a
Fix stack() for non-Tensor types
