### 2023-11-24 d73592e
Consider NaN extrapolations equal
