### 2023-11-24 1de944b
Add argument check in trace_check()
