### 2023-11-24 37f5b77
Fix Solve attrs
