### 2023-11-25 2c23b10
Fix for missing dims in broadcast_op()
