+++ phiml/math/_ops.py
@@ -1103,7 +1103,7 @@ def broadcast_op(operation: Callable,
                 iter_dims.add(tensor._stack_dim.name)
             # --- remove iter_dims for which the sizes vary among tensors ---
             for dim in tuple(iter_dims):
-                sizes = [t.shape.get_size(dim) if dim in t.shape else None for t in tensors]
+                sizes = [t.shape.get_size(dim) for t in tensors if dim in t.shape]
                 if not all(s == sizes[0] for s in sizes[1:]):
                     iter_dims.remove(dim)
     if len(iter_dims) == 0:
