### 2023-11-25 2d25dce
Fix for Tensor-valued offsets in join_index_offsets()
