### 2023-11-25 c98b870
Optimize batched_gather_nd() for PyTorch, Jax
