### 2023-11-25 f40a7fa
Allow None for x0 in map()
