### 2023-11-26 b46948f
Fix Extrapolation.shape
