### 2023-11-26 2934bba
Support flatten() for non-Shaped args
