### 2023-11-26 159df93
always_close() now returns False for incompatible tensors
