### 2023-11-27 2296ea1
Support SciPy linear solves inside jit-compiled functions
* Non-NumPy Preconditioners not yet supported
