### 2023-11-27 6208875
Fix numpy_call()
