### 2023-11-28 eacadb4
Add Backend.get_sparse_format()
