### 2023-11-28 6624b4d
Fix PyTorch is_available()
