### 2023-11-28 a70c465
SciPy preconditioners in JIT mode
