+++ phiml/backend/_linalg.py
@@ -9,7 +9,8 @@ import scipy
 from scipy.sparse import issparse, coo_matrix
 from scipy.sparse.linalg import spsolve, LinearOperator
 
-from ._backend import Backend, SolveResult, List, DType, spatial_derivative_evaluation, combined_dim, choose_backend, TensorType, Preconditioner, ML_LOGGER, convert
+from ._backend import Backend, SolveResult, List, DType, spatial_derivative_evaluation, combined_dim, choose_backend, TensorType, Preconditioner, ML_LOGGER, convert, disassemble_dataclass
+from ._dtype import to_numpy_dtype, combine_types
 from ._numpy_backend import NUMPY
 
 
@@ -266,7 +267,7 @@ def bicg_stab_first_order(b: Backend, lin, y, x0, rtol, atol, max_iter, pre: Opt
     return SolveResult(f"Φ-ML biCG-stab {pre_str(pre)}", x, residual, iterations, function_evaluations, converged, diverged, [""] * batch_size)
 
 
-def scipy_spsolve(b: Backend, method: Union[str, Callable], lin, y, x0, rtol, atol, max_iter, pre: Optional[Preconditioner]) -> SolveResult:
+def scipy_sparse_solve(b: Backend, method: Union[str, Callable], lin, y, x0, rtol, atol, max_iter, pre: Optional[Preconditioner]) -> SolveResult:
     assert max_iter.shape[0] == 1, f"Trajectory recording not supported for scipy_spsolve"
     if method == 'direct' and pre:
         warnings.warn(f"Preconditioner {pre} was computed but is not used by SciPy direct solve.", RuntimeWarning)
@@ -286,25 +287,28 @@ def scipy_spsolve(b: Backend, method: Union[str, Callable], lin, y, x0, rtol, at
     batch_size = b.staticshape(y)[0]
     if not callable(lin):
         assemble_lin, lin_tensors = b.disassemble(lin)
-        def scipy_solve(np_y, np_x0, np_rtol, np_atol, *np_lin_tensors):
-            np_lin = assemble_lin(NUMPY, *np_lin_tensors)
+        assemble_pre, pre_tensors = disassemble_dataclass(pre)
+        def scipy_solve(np_y, np_x0, np_rtol, np_atol, *np_tensors):
+            np_lin = assemble_lin(NUMPY, *np_tensors[:len(lin_tensors)])
+            np_pre = assemble_pre(NUMPY, *np_tensors[len(lin_tensors):])
             if method == 'direct':
                 npr = scipy_direct_linear_solve(NUMPY, np_lin, np_y)
             else:
-                npr = scipy_iterative_sparse_solve(NUMPY, np_lin, np_y, np_x0, np_rtol, np_atol, max_iter, pre, function)
+                npr = scipy_iterative_sparse_solve(NUMPY, np_lin, np_y, np_x0, np_rtol, np_atol, max_iter, np_pre, function)
             return npr.x, npr.residual, npr.iterations, npr.function_evaluations, npr.converged, npr.diverged
     else:
         assert method != 'direct', "scipy direct matrix solve cannot be used in matrix-free mode"
         lin_tensors = []
-        def scipy_solve(np_y, np_x0, np_rtol, np_atol):
-            npr = scipy_iterative_sparse_solve(NUMPY, lin, np_y, np_x0, np_rtol, np_atol, max_iter, pre, function)
+        assemble_pre, pre_tensors = disassemble_dataclass(pre)
+        def scipy_solve(np_y, np_x0, np_rtol, np_atol, *np_pre_tensors):
+            np_pre = assemble_pre(NUMPY, *np_pre_tensors)
+            npr = scipy_iterative_sparse_solve(NUMPY, lin, np_y, np_x0, np_rtol, np_atol, max_iter, np_pre, function)
[…]
