### 2023-11-28 26b1918
Fix dtype conversion
