### 2023-11-28 a982fd3
Fix COO-dense multiplication
