### 2023-11-28 2b1bcba
Allow gather() along non-existent dims
