### 2023-11-28 0c04be9
Tracing fixes, add unit test
