### 2023-11-30 df30015
Support trees in all_available()
