### 2023-12-04 915f8dc
Changes necessary for one sided boundaries - small fixes connected with vector dimension of fields - introduce parameter 'fix_rank_deficiency' in linear call chain
