+++ phiml/math/extrapolation.py
@@ -1744,6 +1744,18 @@ def map(f: Callable[[Extrapolation], Extrapolation], extrapolation):
         return f(extrapolation)
 
 
+def give_leaves(extrapolation: Extrapolation):
+    if isinstance(extrapolation, _MixedExtrapolation):
+        ret = list(sum([give_leaves(ext) for k, ext in extrapolation.ext.items()], []))
+    elif isinstance(extrapolation, _NormalTangentialExtrapolation):
+        ret = list(give_leaves(extrapolation.normal) + give_leaves(extrapolation.tangential))
+    elif isinstance(extrapolation, _ConditionalExtrapolation):
+        ret = list(give_leaves(extrapolation.true_ext) + give_leaves(extrapolation.false_ext))
+    else:
+        ret = [extrapolation]
+    return list(set(ret))
+
+
 def remove_constant_offset(extrapolation):
     """
     Removes all constant offsets from an extrapolation.
