### 2023-12-04 215f2c3
Changes necessary for one sided boundaries - small fixes connected with vector dimension of fields - introduce parameter 'fix_rank_deficiency' in linear call chain
