+++ phiml/backend/_numpy_backend.py
@@ -506,7 +506,7 @@ class NumPyBackend(Backend):
     #             return grads
     #     return gradient
 
-    def linear_solve(self, method: str, lin, y, x0, rtol, atol, max_iter, pre) -> SolveResult:
+    def linear_solve(self, method: str, lin, y, x0, rtol, atol, max_iter, pre, fix_rank_deficiency=False) -> SolveResult:
         if method in ['direct', 'CG-native', 'GMres', 'biCG', 'biCG-stab', 'CGS', 'lGMres', 'minres', 'QMR', 'GCrotMK'] and max_iter.shape[0] == 1:
             from ._linalg import scipy_sparse_solve
             return scipy_sparse_solve(self, method, lin, y, x0, rtol, atol, max_iter, pre)
