### 2023-12-17 d7c110b
Slice padding fix
