### 2023-12-20 ed3af45
iterate() for functions with auxiliary outputs
