+++ tests/commit/math/test__functional.py
@@ -199,6 +199,26 @@ class TestFunctional(TestCase):
         math.assert_close([1, 2, 4], r)
         self.assertEqual(batch(trajectory=2), t.shape)
 
+    def test_iterate_more_outputs(self):
+        def f(x):
+            return x + 1, 1
+
+        x_trj, one_trj = math.iterate(f, spatial(t=10), 0)
+        math.assert_close(x_trj, math.range(spatial(t=11)))
+        math.assert_close(one_trj, 1)
+        final_x, final_one = math.iterate(f, 10, 0)
+        math.assert_close(final_x, 10)
+        math.assert_close(final_one, 1)
+
+    def test_iterate_without_initial(self):
+        def f(x):
+            return x + 1 if x is not None else 0
+
+        x_trj = math.iterate(f, spatial(t=10), None)
+        math.assert_close(x_trj, math.range(spatial(t=10)))
+        final_x = math.iterate(f, 10, None)
+        math.assert_close(final_x, 9)
+
     def test_delayed_decorator(self):
         def f(x, y):
             return x + y
