### 2023-12-20 f45e5d9
Allow slicing with int32
