+++ phiml/math/_shape.py
@@ -1178,12 +1178,9 @@ class Shape:
                     gathered_sizes = [(int(s) if isinstance(s, Tensor) and s.rank == 0 else s) for s in gathered_sizes]
                     result = result.with_sizes(gathered_sizes, keep_item_names=True).without(sel_dim)
             elif isinstance(selection, slice):
-                assert isinstance(selection.step, int) or selection.step is None, f"slice step must be an int or None but got {type(selection.step).__name__}"
-                assert isinstance(selection.start, int) or selection.start is None, f"slice start must be an int or None but got {type(selection.start).__name__}"
-                assert isinstance(selection.stop, int) or selection.stop is None, f"slice stop must be an int or None but got {type(selection.stop).__name__}"
-                step = selection.step or 1
-                start = selection.start if isinstance(selection.start, int) else (0 if step > 0 else self.get_size(sel_dim)-1)
-                stop = selection.stop if isinstance(selection.stop, int) else (self.get_size(sel_dim) if step > 0 else -1)
+                step = int(selection.step) if selection.step is not None else 1
+                start = int(selection.start) if selection.start is not None else (0 if step > 0 else self.get_size(sel_dim)-1)
+                stop = int(selection.stop) if selection.stop is not None else (self.get_size(sel_dim) if step > 0 else -1)
                 if stop < 0 and step > 0:
                     stop += self.get_size(sel_dim)
                     assert stop >= 0
