### 2023-12-20 99d10cf
Allow slicing with int32
