### 2023-12-22 2af5ced
Gradient descent improvements
