### 2023-12-22 2c4cbdc
Fix matrix gradient in linear_solve()
