### 2023-12-24 11ac924
Add argmax(), argmin(), fix pairwise distances
