### 2023-12-24 1cd0c3f
Add argmax(), argmin(), fix pairwise distances
