### 2023-12-24 a4afc38
Add find_closest()
