### 2023-12-24 90d3a9d
Add find_closest()
