### 2023-12-25 b8b5465
Support multi-type Shape.only(), without()
