### 2023-12-25 1b9d3c5
Support multi-type Shape.only(), without()
