### 2023-12-26 0c806f9
Keep dtype when converting compressed sparse to dense
