### 2023-12-26 dfb0332
Add to_format() for sparse/dense conversion
