### 2023-12-26 f5f404c
Add to_format() for sparse/dense conversion
