### 2023-12-26 7a9da09
Fix COO gather for multi-dim indices
