+++ phiml/backend/torch/_torch_backend.py
@@ -809,7 +809,7 @@ class TorchBackend(Backend):
         for b in range(batch_size):
             b_result = []
             for c in range(channels):
-                matrix = torch.sparse_csr_tensor(row_pointers[b], column_indices[b], values[b, :, c], shape, device=values.device)
+                matrix = torch.sparse_csr_tensor(row_pointers[b], column_indices[b], values[b, :, c].contiguous(), shape, device=values.device)
                 b_result.append(torch.sparse.mm(matrix, self.as_tensor(dense[b, :, c, :])))
             result.append(torch.stack(b_result, 1))
         return torch.stack(result)
