### 2023-12-26 709e8c8
Add is_sparse()
