### 2023-12-26 3b446c3
Add is_sparse()
