### 2023-12-26 9fa4207
Add min/max modes to scatter()
