### 2023-12-26 7430ace
Add min/max modes to scatter()
