+++ phiml/backend/tensorflow/_tf_backend.py
@@ -579,14 +579,10 @@ class TFBackend(Backend):
             base_grid, values = self.auto_cast(base_grid, values)
             indices = self.as_tensor(indices)
             batch_size = combined_dim(combined_dim(indices.shape[0], values.shape[0]), base_grid.shape[0])
-            scatter = tf.tensor_scatter_nd_add if mode == 'add' else tf.tensor_scatter_nd_update
-            result = []
-            for b in range(batch_size):
-                b_grid = base_grid[b, ...]
-                b_indices = indices[min(b, indices.shape[0] - 1), ...]
-                b_values = values[min(b, values.shape[0] - 1), ...]
-                result.append(scatter(b_grid, b_indices, b_values))
-            return self.stack(result, axis=0)
+            scatter = {'add': tf.tensor_scatter_nd_add, 'update': tf.tensor_scatter_nd_update, 'max': tf.tensor_scatter_nd_max, 'min': tf.tensor_scatter_nd_min}[mode]
+            def scatter_single(b_grid, b_indices, b_values):
+                return scatter(b_grid, b_indices, b_values)
+            return self.vectorized_call(scatter_single, base_grid, indices, values, output_dtypes=self.dtype(base_grid))
 
     def histogram1d(self, values, weights, bin_edges):
         with self._device_for(values, weights, bin_edges):
