### 2023-12-26 4f8284e
min, max, any for sparse matrices
