### 2023-12-26 dc3918b
Sparse padding for dense tensors
