### 2023-12-26 bde627e
sparse gather for compressed matrices
