### 2023-12-26 6cd2687
argmax for sparse tensors
