### 2023-12-26 49389ca
Fix nonzero for sparse tensors
