### 2023-12-26 07292db
Add 'default' to scatter() to prevent type upgrades
