### 2023-12-30 32511aa
Fix for TensorStack.__pack_dims__()
