### 2023-12-30 98bc468
Fix Solve.__eq__
