### 2023-12-30 9242b3a
minimize() for non-uniform tensors
