### 2023-12-30 2ba0a3d
minimize() for non-uniform tensors
