+++ phiml/math/_optimize.py
@@ -10,7 +10,7 @@ from ..backend import get_precision, NUMPY, Backend
 from ..backend._backend import SolveResult, ML_LOGGER, default_backend, convert, Preconditioner, choose_backend
 from ..backend._linalg import IncompleteLU, incomplete_lu_dense, incomplete_lu_coo, coarse_explicit_preconditioner_coo
 from ._shape import EMPTY_SHAPE, Shape, merge_shapes, batch, non_batch, shape, dual, channel, non_dual, instance, spatial
-from ._magic_ops import stack, copy_with, rename_dims, unpack_dim
+from ._magic_ops import stack, copy_with, rename_dims, unpack_dim, unstack, expand
 from ._sparse import native_matrix, SparseCoordinateTensor, CompressedSparseMatrix, stored_values, is_sparse
 from ._tensors import Tensor, disassemble_tree, assemble_tree, wrap, cached, NativeTensor, layout
 from . import _ops as math
@@ -369,20 +369,36 @@ def minimize(f: Callable[[X], Y], solve: Solve[X, Y]) -> X:
     backend = choose_backend_t(*x0_tensors, prefer_default=True)
     batch_dims = merge_shapes(*[t.shape for t in x0_tensors]).batch
     x0_natives = []
+    x0_native_shapes = []
     for t in x0_tensors:
-        t._expand()
-        assert t.shape.is_uniform
-        x0_natives.append(reshaped_native(t, [batch_dims, t.shape.non_batch]))
+        t = cached(t)
+        if t.shape.is_uniform:
+            x0_natives.append(reshaped_native(t, [batch_dims, t.shape.non_batch]))
+            x0_native_shapes.append(t.shape.non_batch)
+        else:
+            for ut in unstack(t, t.shape.shape.without('dims')):
+                x0_natives.append(reshaped_native(ut, [batch_dims, ut.shape.non_batch]))
+                x0_native_shapes.append(ut.shape.non_batch)
     x0_flat = backend.concat(x0_natives, -1)
 
     def unflatten_assemble(x_flat, additional_dims: Shape = EMPTY_SHAPE, convert=True):
+        partial_tensors = []
         i = 0
-        x_tensors = []
-        for x0_native, x0_tensor in zip(x0_natives, x0_tensors):
-            vol = backend.shape(x0_native)[-1]
+        for x0_native, t_shape in zip(x0_natives, x0_native_shapes):
+            vol = backend.staticshape(x0_native)[-1]
             flat_native = x_flat[..., i:i + vol]
-            x_tensors.append(reshaped_tensor(flat_native, [*additional_dims, batch_dims, x0_tensor.shape.non_batch], convert=convert))
+            partial_tensor = reshaped_tensor(flat_native, [*additional_dims, batch_dims, t_shape], convert=convert)
+            partial_tensors.append(partial_tensor)
             i += vol
+        # --- assemble non-uniform tensors ---
+        x_tensors = []
+        for t in x0_tensors:
+            if t.shape.is_uniform:
+                x_tensors.append(partial_tensors.pop(0))
+            else:
+                stack_dims = t.shape.shape.without('dims')
+                x_tensors.append(stack(partial_tensors[:stack_dims.volume], stack_dims))
+                partial_tensors = partial_tensors[stack_dims.volume:]
         x = assemble_tree(x0_nest, x_tensors)
         return x
 
@@ -394,17 +410,22 @@ def minimize(f: Callable[[X], Y], solve: Solve[X, Y]) -> X:
             y = f(x)
         _, y_tensors = disassemble_tree(y)
[…]
