### 2023-12-30 93ca00e
Add equal()
