### 2024-01-01 c0a31c3
Decompress matrix if backend does not support compressed type
