### 2024-01-01 79b1ea9
Store Tensor.__init__() stack trace when debug checks are enabled
