### 2024-01-01 e38c0af
Store Tensor.__init__() stack trace when debug checks are enabled
