### 2024-01-01 5159339
Add troubleshoot cache_all_tensors()
