### 2024-01-01 068e2bd
Convert x0 in solve_linear()
