+++ phiml/math/_optimize.py
@@ -603,7 +603,8 @@ def solve_linear(f: Union[Callable[[X], Y], Tensor],
 
     if isinstance(f, Tensor) or (isinstance(f, LinearFunction) and prefer_explicit):  # Matrix solve
         if isinstance(f, LinearFunction):
-            matrix, bias = f.sparse_matrix_and_bias(solve.x0, *f_args, **f_kwargs)
+            x0 = math.convert(solve.x0)
+            matrix, bias = f.sparse_matrix_and_bias(x0, *f_args, **f_kwargs)
         else:
             matrix = f
             bias = 0
