### 2024-01-03 30757a4
Retain item names in 2D cross_product()
