### 2024-01-05 fb03697
Add erf(x)
