### 2024-01-06 396924b
Fix stack for Tensor + Layout
