### 2024-01-06 d36f287
Fix batched gradient descent
