+++ phiml/math/_tensors.py
@@ -10,7 +10,7 @@ from typing import Tuple, Callable, List
 import numpy
 import numpy as np
 
-from ._magic_ops import PhiTreeNodeType, variable_attributes, copy_with, stack, pack_dims, expand
+from ._magic_ops import PhiTreeNodeType, variable_attributes, copy_with, stack, pack_dims, expand, slice_, flatten
 from ._shape import (Shape,
                      CHANNEL_DIM, BATCH_DIM, SPATIAL_DIM, EMPTY_SHAPE,
                      parse_dim_order, shape_stack, merge_shapes, channel, concat_shapes, primal,
@@ -18,7 +18,7 @@ from ._shape import (Shape,
 from ..backend import NoBackendFound, choose_backend, BACKENDS, get_precision, default_backend, convert as convert_, \
     Backend, ComputeDevice, OBJECTS
 from ..backend._dtype import DType, combine_types
-from .magic import BoundDim, PhiTreeNode, slicing_dict
+from .magic import BoundDim, PhiTreeNode, slicing_dict, Shaped
 from .magic import Shapable
 
 
@@ -942,10 +942,28 @@ class Layout(Tensor):
     The PyTree may be deeper but only the outer `shape.rank` levels are represented as a tensor.
     """
 
-    def __init__(self, obj, shape: Shape):
+    def __init__(self, obj, stack_dim: Shape):
         super().__init__()
         self._obj = obj
-        self._shape = shape
+        obj_shapes = Layout._recursive_get_shapes(obj, stack_dim)
+        self._shape = shape_stack(stack_dim, *obj_shapes, stack_dim_first=True)
+        self._stack_dim = stack_dim
+        if DEBUG_CHECKS:
+            if self._stack_dim:
+                assert stack_dim == self._shape[:stack_dim.rank]
+            elif isinstance(obj, Shapable) and obj is not None:
+                warnings.warn(f"Empty stack_dim for Layout with value {obj}")
+
+    @staticmethod
+    def _recursive_get_shapes(obj, s: Shape) -> Tuple[Shape]:
+        if not s:
+            return shape(obj, allow_unshaped=True),
+        elif isinstance(obj, (tuple, list)):
+            return sum([Layout._recursive_get_shapes(o, s.after_gather({s.names[0]: i})) for i, o in enumerate(obj)], ())
+        elif isinstance(obj, dict):
+            return sum([Layout._recursive_get_shapes(v, s.after_gather({s.names[0]: i})) for i, (k, v) in enumerate(obj.items())], ())
+        obj_shape = shape(obj, allow_unshaped=True)
+        return (obj_shape,) * s.volume
 
     @property
     def shape(self) -> Shape:
@@ -968,7 +986,7 @@ class Layout(Tensor):
 
     def native(self, order: Union[str, tuple, list, Shape] = None, singleton_for_const=False):
         order = parse_dim_order(order)
-        assert order is None or order == self._shape.names, "Layout.native() does not allow for changing the dimension order"
+        assert order is None or order == self._stack_dim.names, "Layout.native() does not allow for changing the dimension order"
         return self._obj
 
     def numpy(self, order: Union[str, tuple, list, Shape] = None) -> np.ndarray:
@@ -976,10 +994,20 @@ class Layout(Tensor):
         return numpy.asarray(native)
 
     def _getitem(self, selection: dict) -> 'Tensor':
-        selection_list = [selection.get(dim, None) for dim in self._shape.names]
-        native = self._getitem_recursive(self._obj, tuple(selection_list))
-        new_shape = self._shape.after_gather(selection)
-        return Layout(native, new_shape)
+        selection_list = [selection.get(dim, None) for dim in self._stack_dim.names]
[…]
