### 2024-01-07 cc166ae
math.slice() for primitives
