### 2024-01-07 bc6c871
Simplify result of shape_stack()
