### 2024-01-07 53dc104
Simplify result of shape_stack()
