### 2024-01-07 6036068
Support mapping object dimensions by passing dims='object'
