### 2024-01-07 e9b160c
Add allow_unshaped to shape()
