### 2024-01-07 4c14b6f
Fix PyTorch np.bool_ recognition
