### 2024-01-07 62d5e8c
Fix PyTorch np.bool_ recognition
