### 2024-01-10 #6171 @mstimberg
:wave: Hi everyone, and a Happy New Year :sparkles: !

@wandeln and @chaoming0625 I see that you checked off most/all the boxes in the checklist. Are you still planning on commenting on any specifics and/or suggest any changes to the software or paper? Or would you recommend it for acceptance in JOSS in its current state? Thanks :pray:
