### 2024-01-11 5b30e1b
dot() for non-uniform tensors
