### 2024-01-14 4106d59
scatter mode=prod/any/all
* Also support passing reduction function
