### 2024-01-14 ffe2df3
scatter mode=prod/any/all
* Also support passing reduction function
