### 2024-01-14 485c2ad
transpose() for Shapable
* Add unit tests
