### 2024-01-14 acd61ad
Fix _expand() for sparse tensors
