### 2024-01-14 8b79bc1
Fix _expand() for sparse tensors
