### 2024-01-14 2f9dbf4
Remove Tensor._expand()
* Now fully replaced by cached(Tensor)
