### 2024-01-14 8411220
Implement __eq__ for _SymmetricGradientExtrapolation
