### 2024-01-14 699571e
Add cache to disassemble_tree()
