+++ phiml/_troubleshoot.py
@@ -141,7 +141,7 @@ def plot_solves():
             for i, result in enumerate(solves):
                 assert isinstance(result, math.SolveInfo)
                 from .math._tensors import disassemble_tree
-                _, (residual,) = disassemble_tree(result.residual)
+                _, (residual,) = disassemble_tree(result.residual, cache=False)
                 residual_mse = math.mean(math.sqrt(math.sum(residual ** 2)), residual.shape.without('trajectory'))
                 residual_mse_max = math.max(math.sqrt(math.sum(residual ** 2)), residual.shape.without('trajectory'))
                 # residual_mean = math.mean(math.abs(residual), residual.shape.without('trajectory'))
