### 2024-01-14 ce88b44
Add cache to disassemble_tree()
