+++ phiml/math/_functional.py
@@ -106,7 +106,7 @@ def key_from_args(args: tuple, kwargs: Dict[str, Any], parameters: Tuple[str, ..
             if param in kwargs:
                 aux_kwargs[param] = kwargs[param]
                 del kwargs[param]
-    tree, tensors = disassemble_tree(kwargs)
+    tree, tensors = disassemble_tree(kwargs, cache=cache)
     tracing = not math.all_available(*tensors)
     backend = math.choose_backend_t(*tensors)
     natives, shapes, specs = disassemble_tensors(tensors, expand=cache)
@@ -173,7 +173,7 @@ class JitFunction:
                 in_tensors = assemble_tensors(natives, in_key.specs)
                 kwargs = assemble_tree(in_key.tree, in_tensors)
                 f_output = self.f(**kwargs, **in_key.auxiliary_kwargs)  # Tensor or tuple/list of Tensors
-                tree, out_tensors = disassemble_tree((f_output, self._extract_tensors))
+                tree, out_tensors = disassemble_tree((f_output, self._extract_tensors), cache=True)
                 result_natives, result_shapes, specs = disassemble_tensors(out_tensors, expand=True)
                 self.recorded_mappings[in_key] = SignatureKey(jit_f_native, tree, result_shapes, specs, in_key.backend, in_key.tracing)
             finally:
@@ -288,21 +288,6 @@ class LinearFunction(Generic[X, Y], Callable[[X], Y]):
         self.forget_traces = forget_traces
         self.matrices_and_biases: Dict[SignatureKey, Tuple[SparseCoordinateTensor, Tensor, Tuple]] = {}
         self.nl_jit = JitFunction(f, self.auxiliary_args, forget_traces)  # for backends that do not support sparse matrices
-
-    # def _trace(self, in_key: SignatureKey, prefer_numpy: bool) -> 'ShiftLinTracer':
-    #     assert in_key.shapes[0].is_uniform, f"math.jit_compile_linear() only supports uniform tensors for function input and output but input shape was {in_key.shapes[0]}"
-    #     with NUMPY if prefer_numpy else in_key.backend:
-    #         x = math.ones(in_key.shapes[0])
-    #         tracer = ShiftLinTracer(x, {EMPTY_SHAPE: math.ones()}, x.shape, math.zeros(x.shape))
-    #     _TRACING_JIT.append(self)
-    #     x_kwargs = assemble_tree(in_key.tree, [tracer])
-    #     result = self.f(**x_kwargs, **in_key.auxiliary_kwargs)
-    #     _, result_tensors = disassemble_tree(result)
-    #     assert len(result_tensors) == 1, f"Linear function must return a single Tensor or tensor-like but got {result}"
-    #     result_tensor = result_tensors[0]
-    #     assert isinstance(result_tensor, ShiftLinTracer), f"Tracing linear function '{f_name(self.f)}' failed. Make sure only linear operations are used."
-    #     assert _TRACING_JIT.pop(-1) is self
-    #     return result_tensor
 
     def _get_or_trace(self, key: SignatureKey, args: tuple, f_kwargs: dict):
         if not key.tracing and key in self.matrices_and_biases:
@@ -502,7 +487,7 @@ class GradientFunction:
                 loss_shape = in_key.backend.staticshape(loss_native)
                 assert len(
                     loss_shape) == 0, f"Only scalar losses are allowed when returning a native tensor but {f_name(self.f)} returned {type(loss_native).__name__} of shape {loss_shape}. For higher-dimensional values, use Φ-ML tensors instead."
-            nest, out_tensors = disassemble_tree(result)
+            nest, out_tensors = disassemble_tree(result, cache=True)
[…]
