### 2024-01-15 #6171 @chaoming0625
Dear Philipp,

This is an excellent work that introduces a systematic array/tensor operation paradigm based on explicitly named dimensions. However, I have some questions:

1. The related work section could be expanded. Assigning names or labels to tensor dimensions to increase readability and reliability has been explored in various libraries like `xarray` (2017), `labeled_tensor` in TensorFlow, `namedtensor` in PyTorch, and recently `einops` package.
2. The limitations of this approach should also be discussed, such as the cognitive load for users to remember many axis labels, potential speed impacts of axis ordering (also see the minor comments), errors from automatic reshaping if axis mistakes are not caught, and whether this paradigm can be widely applied in physical sciences or only a subset where explicit labels are common. Named tensors have been proposed for years with little adoption in ML - is physical science a more fitting domain?
3. According to the paper's statements, it seems only support custom CUDA operations in TensorFlow. Are other frameworks like PyTorch supported?
4. The paper title may be too broad given the limitations of named tensors. ``Science`` covers an extremely wide range of domains, while this approach may not be applicable or have adoption issues in many areas. Perhaps the application domain in the title could be narrowed down.

Minor comments:

1. There appear to be some bugs or compatibility issues in the framework - see the documentation at https://tum-pbs.github.io/PhiML/Linear_Solves.html.
2. Performance compared to raw PyTorch/JAX should be benchmarked. The paper claims execution speed is guaranteed, but no data demonstrates Python overhead with these wrappers vs. PyTorch's impressive native performance.

Finally, thank you for the submission. I hope these suggestions are helpful for improving the paper. Please let me know if you would like me to clarify or expand on any point.

Best,

Chaoming Wang
