### 2024-01-16 afe6663
Add named dims references to paper
