+++ paper.bib
@@ -288,3 +288,32 @@
   version = {0.2.5},
   year = {2018},
 }
+
+
+@article{xarray2017,
+  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  author    = {Hoyer, S. and J. Hamman},
+  journal   = {Journal of Open Research Software},
+  volume    = {5},
+  number    = {1},
+  year      = {2017},
+  publisher = {Ubiquity Press},
+  doi       = {10.5334/jors.148},
+  url       = {https://doi.org/10.5334/jors.148}
+}
+
+@misc{NamedTensor,
+  author = {Harvard NLP},
+  title = {NamedTensor},
+  howpublished = {\url{https://github.com/harvardnlp/NamedTensor}},
+  year = {2019},
+  note = {Accessed on 16 Jan 2024}
+}
+
+@misc{Einops2018,
+  author = {Aleksey Rogozhnikov},
+  title = {Einops},
+  howpublished = {\url{https://github.com/arogozhnikov/einops}},
+  year = {2018},
+  note = {Accessed on 16 Jan 2024}
+}
