### 2024-01-16 99cc716
Improved gather/scatter checks
