### 2024-01-16 84b5c7a
Auto-convert NumPy scalars to primitives in functional math
This avoids NumPy conversion errors. This only affects the case that non-Tensors are passed.
