+++ phiml/math/_tensors.py
@@ -1927,6 +1927,8 @@ def assemble_tree(obj: PhiTreeNodeType, values: List[Tensor]) -> PhiTreeNodeType
     elif obj is NATIVE_TENSOR:
         value = values.pop(0)
         assert isinstance(value, NativeTensor), f"Failed to assemble tree structure. Encountered {value}"
+        if isinstance(value._native, np.ndarray) and value.shape == EMPTY_SHAPE:  # this can be represented as a Python scalar, which leads to less conversion errors
+            return value._native.item()
         return value._native
     elif obj is None:
         value = values.pop(0)
