### 2024-01-16 #6171 @holl-
Hi @chaoming0625,

Thanks for your great comments! I've added the proposed references (except for labeled_tensor which doesn't seem to exist anymore) and fixed the error in the documentation.

Custom CUDA kernels are currently only available for TensorFlow. However, these are optional and only increase performance for certain operations which are not natively implemented in TensorFlow. We have considered adding custom PyTorch operations, but this would only lead to minor gains in performance.
We found that a much larger gain in performance can be achieved by switching to Jax with JIT. This level of performance is hard to beat, even with custom CUDA kernels.

By *science-oriented*, we don't mean to say that all of science should use this library. Instead, PhiML focusses scientific applications rather than pure machine learning. For applications that focus more on network architectures and can work with supervised learning, there is little benefit to using PhiML. This library shines when it comes to differentiable simulations, optimization with analytic derivatives, and integrating machine learning tools into scientific applications.
PhiML covers a similar range as NumPy/SciPy, mapping a large part of their APIs. The core features, such as named dimensions and efficient (sparse) linear solves, can be applied on a wide variety of scientific applications.

Automatic reshaping does have an impact on performance, but only when running the Python code as-is. Once JIT-compiled, these overheads are fully optimized out. As for performance benchmarks, I'm planning to write a simple simulation in PhiML and the same in PyTorch/TensorFlow/Jax natively to compare the runtime with JIT-compilation. Is this what you meant?

Unlike with related libraries, our named dimension implementation is only one part of the puzzle. The arguably bigger impact is achieved by *dimension types*. In fact, all library functions are agnostic to the dimension names and select the dimensions to operate on based on their types. The same is recommended for all user-written library/helper functions. Dimension names should only be assigned at the highest level by the user, and should ideally be fully contained in one file. Remembering a handful of dimension names seems no more difficult than remembering the dimension order in another library, something that is not necessary in PhiML.

We have applied this paradigm to many different kinds of tasks, from custom neural network libraries to PDEs to rigid body simulations to curve fitting problems. Despite there being only a handful of dimension types, the system has scaled remarkably well to all settings. Dimension names are especially useful in scientific tasks as a wider range of tensor functions is used there compared to pure machine learning applications.
I'll add an example to show how dimension names prevent errors by only allowing correct / sensible operations.

Best,
Philipp
