### 2024-01-17 #6171 @holl-
@chaoming0625 I've uploaded a preliminary [performance comparison notebook](https://tum-pbs.github.io/PhiML/Performance.html), comparing Φ-ML vs native performance.
