### 2024-01-17 982fbf5
Fix padding=0 in shift()
