### 2024-01-17 #6171 @holl-
@chaoming0625 I've also uploaded a preliminary [notebook](https://tum-pbs.github.io/PhiML/Dimension_Names_Types.html) showcasing the advantages of named and typed dimensions. I was again surprised by how difficult many things are in PyTorch.
