### 2024-01-19 b0719b3
Fix expand non-uniform
