### 2024-01-19 9948e8f
Use non-batch dims in tensor[bool]
