### 2024-01-20 152701b
Ignore extra batch dimensions in minimize() if they have volume 1
