### 2024-01-21 #6171 @chaoming0625
Dear @holl- ,

Thank you for addressing most of my concerns. One additional thought is still on the `science-oriented` aspect.

Based on my understanding, numerical computing frameworks like NumPy, PyTorch, and JAX have demonstrated significant impact in scientific computing (*[Nature 585, 357–362, 2020](https://www.nature.com/articles/s41586-020-2649-2)*; *[Nature Methods, 17, 261–272](https://www.nature.com/articles/s41592-019-0686-2)*). These frameworks are specifically designed for scientific purposes and have become ubiquitous across scientific disciplines.

Considering this, it's debatable to claim $\Phi_\textrm{ML}$ is uniquely "science-oriented" relative to previous solutions.  The current phrasing implies past frameworks were not designed for science, which is inaccurate. Perhaps it would be better to highlight $\Phi_\textrm{ML}$'s specific features that make it well-suited and user-friendly for scientists, rather than positioning it as the first science-oriented framework, given the current paper title.

Best,

Chaoming
