### 2024-01-21 #6171 @chaoming0625
The current problem statement reflects your perspective, but it does not imply that existing frameworks are incapable of addressing these needs (L23-L31). On the contrary, current math libraries have functionalities to tackle these challenges, although their solutions may not be as seamless or user-friendly.

For instance, it is not clear how the first problem can be solved (L24) and why other libraries cannot address it. Additionally, the third problem (L29) related to sparsity has also been well-supported in all the listed libraries. Moreover, the last problem (L31) has been well-addressed in existing math libraries. Notably, JAX allows computation based on x64 by simply setting ``jax.config.update("jax_enable_x64", True)``.

Considering these points, the current paper may overstate the limitations of the existing options.

I believe the current paper should highlight the second problem (L27), in which ΦML provides simpler and more intuitive solutions compared to existing frameworks, rather than claiming to address previously unmet needs.
