### 2024-01-22 324a52d
Improved Shape error message
