### 2024-01-24 e2d77ea
Add incomplete_gamma()
