### 2024-01-25 #6171 @holl-
@chaoming0625 Thank you for your clarifications. I agree that the point made in line 24 is misleading. While by itself, it's correct, it is not connected to a limitation of existing frameworks like the other points. I'll change or remove it.

For the three other points, I'll work on a new notebook to clearly show how PhiML improves upon the existing libraries. To clarify, it's not that other libraries are incapable of achieving certain results, it's that getting there may be tedious. For example, implementing a custom gradient operation is possible with all frameworks but having to do it is inconvenient.

I'll update you as soon as I have the notebook ready!
