### 2024-01-29 #6171 @mstimberg
@wandeln could you please get back to me about the state of your review :point_down: ?
> I see that you checked off most/all the boxes in the checklist. Are you still planning on commenting on any specifics and/or suggest any changes to the software or paper? Or would you recommend it for acceptance in JOSS in its current state? Thanks 🙏
