### 2024-01-31 32cc45e
Keep item names in concat() for 0-length tensors
