### 2024-01-31 2abaeda
Improved error message for expand tensor
