### 2024-01-31 37651ea
Allow empty vec()
