### 2024-01-31 747b9d4
Support Shape[Shape] to select by dim names
