### 2024-01-31 65dce28
Fix test_get_backend()
