### 2024-02-05 #6171 @wandeln
Hi @mstimberg, I tested Phi-ML with Pytorch and Tensorflow on my machine and it seems to work as described. I think it is a useful library if you want to develop code that is agnostic wrt the underlying DL framework and since it was developed explicitly to form the backbone of Phi-flow, I see it's purpose for scientific applications. Unfortunately, I have not a lot of experience with platform-agnostic DL frameworks so I would score the confidence of my rating rather low (especially regarding related work).
