### 2024-02-07 #6171 @chaoming0625
Hello @mstimberg ,

I hope this message finds you well. I wanted to take a moment to summarize my thoughts regarding the software. In general, I find ΦML to be a very valuable tool as it brings systematic named tensors to the forefront of mainstream computing frameworks such as NumPy, PyTorch, TensorFlow, and JAX. However, I believe it would greatly benefit from providing more thorough justifications for its motivations, contributions, and innovative aspects. The current wording appears somewhat misleading, particularly in the section spanning lines 24 to 41. Once the authors have made the necessary revisions, I highly recommend considering publishing the paper on JOSS.

Thank you for your attention to these suggestions.

Best regards,

Chaoming Wang
