### 2024-02-08 #6171 @holl-
Hello @chaoming0625 and @mstimberg

Thank you again for your helpful feedback! I have updated the passage you pointed out and uploaded [a new notebook](https://tum-pbs.github.io/PhiML/Advantages_Data_Types.html) showing the advantages of our precision management system compared to the existing ML libraries. Along with the other notebooks I uploaded, this should help motivate our library and design choices.

Cheers,
Philipp
