### 2024-02-08 741e7fc
Ignore transpose() if order is correct
