### 2024-02-08 2249df6
Support indices in Shape.only()
