### 2024-02-08 f0bcd77
Use value attributes for optimization
