### 2024-02-08 2e15b84
Use value attributes for optimization
