### 2024-02-08 1f8c17c
Simplify stop_gradient()
