### 2024-02-10 fe20569
Use value_attrs in functional math
