+++ phiml/math/_functional.py
@@ -8,7 +8,7 @@ from typing import Tuple, Callable, Dict, Generic, List, TypeVar, Any, Set, Unio
 import numpy as np
 
 from . import _ops as math, all_available
-from ._magic_ops import stack, slice_
+from ._magic_ops import stack, slice_, value_attributes
 from ._shape import EMPTY_SHAPE, Shape, spatial, instance, batch, channel, merge_shapes, DimFilter, shape
 from ._sparse import SparseCoordinateTensor
 from ._tensors import Tensor, disassemble_tree, assemble_tree, disassemble_tensors, assemble_tensors, variable_attributes, wrap, specs_equal, equality_by_shape_and_value, object_dims
@@ -98,7 +98,12 @@ def match_output_signature(new_in: SignatureKey, recorded_mappings: Dict[Signatu
                        f"Registered transforms:\n{transforms_str}")  # KeyError does not support \n
 
 
-def key_from_args(args: tuple, kwargs: Dict[str, Any], parameters: Tuple[str, ...], cache=False, aux: Set[str] = ()) -> Tuple[SignatureKey, List[Tensor], tuple, Dict[str, Any]]:
+def key_from_args(args: tuple,
+                  kwargs: Dict[str, Any],
+                  parameters: Tuple[str, ...],
+                  cache=False,
+                  aux: Set[str] = (),
+                  attr_type=variable_attributes) -> Tuple[SignatureKey, List[Tensor], tuple, Dict[str, Any]]:
     kwargs = {**kwargs, **{parameters[i]: v for i, v in enumerate(args)}}
     aux_kwargs = {}
     if aux:
@@ -106,7 +111,7 @@ def key_from_args(args: tuple, kwargs: Dict[str, Any], parameters: Tuple[str, ..
             if param in kwargs:
                 aux_kwargs[param] = kwargs[param]
                 del kwargs[param]
-    tree, tensors = disassemble_tree(kwargs, cache=cache)
+    tree, tensors = disassemble_tree(kwargs, cache=cache, attr_type=attr_type)
     tracing = not math.all_available(*tensors)
     backend = math.choose_backend_t(*tensors)
     natives, shapes, specs = disassemble_tensors(tensors, expand=cache)
@@ -476,7 +481,7 @@ class GradientFunction:
         def f_native(*natives):
             ML_LOGGER.debug(f"Φ-ML-grad: Evaluating gradient of {f_name(self.f)}")
             in_tensors = assemble_tensors(natives, in_key.specs)
-            kwargs = assemble_tree(in_key.tree, in_tensors)
+            kwargs = assemble_tree(in_key.tree, in_tensors, attr_type=value_attributes)
             with functional_derivative_evaluation(order=1):
                 result = self.f(**kwargs)  # Tensor or tuple/list of Tensors
             loss = result[0] if isinstance(result, (tuple, list)) else result
@@ -488,7 +493,7 @@ class GradientFunction:
                 loss_shape = in_key.backend.staticshape(loss_native)
                 assert len(
                     loss_shape) == 0, f"Only scalar losses are allowed when returning a native tensor but {f_name(self.f)} returned {type(loss_native).__name__} of shape {loss_shape}. For higher-dimensional values, use Φ-ML tensors instead."
-            nest, out_tensors = disassemble_tree(result, cache=True)
+            nest, out_tensors = disassemble_tree(result, cache=True, attr_type=value_attributes)
             result_natives, result_shapes, specs = disassemble_tensors(out_tensors, expand=True)
             self.recorded_mappings[in_key] = SignatureKey(f_native, nest, result_shapes, specs, in_key.backend, in_key.tracing)
             return loss_native, result_natives
@@ -499,7 +504,7 @@ class GradientFunction:
[…]
