### 2024-02-14 #6171 @mstimberg
Hi @chaoming0625 did you have a chance to look at the updated manuscript and the notebook @holl- linked to in [this comment](https://github.com/openjournals/joss-reviews/issues/6171#issuecomment-1934530555)? Please let me know whether the changes/additions address your concerns :pray: Thanks!
