### 2024-02-15 #6171 @chaoming0625
Thanks for the remind @mstimberg , and thanks for the wonderful work @holl- .

However, more comments or suggestions are (can be ignored):

No more notebooks for demonstrating the capability of the software are needed. The software has done a wonderful work, while, from my own perspective, the presentation of the work is the place to improve. This work should emphase its intrinsic novelty, rather than some points have been satisfied or unsolved. For example,

- ``A high numerical accuracy is desirable, often requiring 64-bit floating point calculations.`` As have been pointed, 64-bit computation has been intrinsicly supported in all existing frameworks. However, the author presents examples in [a notebook](https://tum-pbs.github.io/PhiML/Advantages_Data_Types.html) demonstrating the power of a precision context manager for controlling the *precision conversions*. Moreover, similar context manager has also existed. For instance, jax has  ``jax.experimental.enable_x64`` manager.
- ``may require implementing custom operators, since many common functions like sparse-sparse matrix multiplication, periodic padding or triangular solves are not available in all libraries.``  I do not see any contributions to solve problems like sparse-sparse matrix multiplication in this work. I do not know whether it is suitable to add a discussion section in a JOSS paper. I understand sparsity is the intrinsic property of the physical influence, however, the work does not contribute to here.
- ``Support for Jax, PyTorch, TensorFlow & NumPy`` has a different title levels compared to the others.
- ``"ΦML: intuitive and reliable math library with dimension types for ..."`` or others like this may be more better than the current title? If you say ``ΦML: A Science-oriented Math and Neural Network Library``, I do not know what you say. And the work may be drowned out among JAX, Torch, TF, or other widely accepted frameworks.
- You say ``science-oriented``, however, the software is called Φ-``ML``? what does ``ML`` mean?

I the editor were to accept this paper directly, I would also support.

Thanks again.
