### 2024-02-15 #6171 @holl-
@chaoming0625,

Thank you for your support and for taking the time to read over the additions!

I agree with your assessment of the paper title. I now changed it to `Φ-ML: Intuitive scientific computing with dimension types for Jax, PyTorch, TensorFlow & NumPy`. I hope this expresses that the target applications of the library are simulations and inverse problems from science and engineering, and that integration with ML methods is a key aspect.

Actually matrix-matrix multiplications are on my to-do list 😅 but I've changed the text now.

> `Support for Jax, PyTorch, TensorFlow & NumPy` has a different title levels compared to the others.

Do you mean the order in which the libraries are mentioned is not consistent or that they should be ordered differently?

Cheers!
