### 2024-02-15 #6171 @holl-
@chaoming0625 I got the sparse matrix-matrix multiplication working for all backends now 😁. May not be as fast as native CUDA implementations but it works. But anyway, this was just an example to illustrate that the ML libraries primarily target core ML applications, such as CV, LLMs, etc. There are many other examples, e.g. PyTorch didn't even have complex numbers at launch.
