### 2024-02-16 #6171 @chaoming0625
> @chaoming0625 I got the sparse matrix-matrix multiplication working for all backends now 😁. May not be as fast as native CUDA implementations but it works. But anyway, this was just an example to illustrate that the ML libraries primarily target core ML applications, such as CV, LLMs, etc. There are many other examples, e.g. PyTorch didn't even have complex numbers at launch.

This does not address my concern. On the contrary, this increases my concerns. If you want to add this functionality, please show me its demonstration and performance. Only implementing sparse-sparse matmul using APIs of the underlying framework has no novelty. Without your sparse-sparse, users can implement them from scratch. So, what's the advantage of using your implementation? Show me, and demonstrate it in the paper.

The current paper has a very bad writing. No demonstrations, no data, and no evaluations. Please learn from previously published JOSS papers.
