### 2024-02-16 #6171 @holl-
@chaoming0625, I'm not sure that what you are requesting should go into the paper. Citing the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html):

> JOSS papers are only expected to contain a limited set of metadata (see example below), a Statement of need, Summary, Acknowledgements, and References sections.

It is my understanding that the project documentation is the place to highlight novelty and demonstrate capabilities, not the paper.
@mstimberg Could you clarify this point?

> This is the current title hierarchy.

Ok, I see the confusion. This may be a bug in your display software. The reason is that I'm using `###` for the subsections. @mstimberg Is this a problem? I can change it to bold text or `##`, but `###` gave the best visual results.
