### 2024-02-16 #6171 @mstimberg
Dear @chaoming0625, thanks again for your detailed review. As described in the guidelines @holl- linked to above, JOSS papers are quite a bit different from papers in other journals (and most of the review is usually focussed on the software and its documentation rather than on the paper). In particular, the paper is supposed to be at around 250–1000 words, which are already greatly exceeded by the current manuscript. For this reason, I don't think it will be possible (or necessary) to have extensive evaluations and/or demonstrations as part of the manuscript (our guidelines state: " a “full length” paper is not permitted"). My suggestion would instead be to significantly shorten the current Design Principles and Major Design Decisions sections (e.g. the subsections in Design Principles could be collapsed into a list with a single sentence explaining each; the Major Design Decisions could be made much shorter and ideally be integrated into the "Statement of Need", i.e. in order to illustrate the difference to other packages). Note that of course I am not saying that the current text is not useful – you might consider putting its unabridged version into the documentation, for example.
What is missing, though, and would at least go a bit into the direction that @chaoming0625 is suggesting, is "Mention (if applicable) a representative set of past or ongoing research projects using the software and recent scholarly publications enabled by it.". If you have references or examples of things that were done with Φ-ML, this would reinforce the point that it is a useful tool.

> Ok, I see the confusion. The reason is that I'm using `###` for the subsections. @mstimberg Is this a problem? I can change it to bold text or `##`, but `###` gave the best visual results.

This is a due to limitations in LaTeX, which does not support `\subsubsection` without a `\subsection`. If you need `\subsection` (but see my suggestions for reorganization above), please use a continuous hierarchy (i.e. `#` followed by `##`, etc.).

Please let me know whether all this works for you, and don't hesitate to ask for further guidance. @chaoming0625 I hope this works for you as well?
