### 2024-02-16 #6171 @holl-
Dear @mstimberg, thank you for your input! I'll shorten the paper and link to the publications that already build on this library.
