### 2024-02-17 d8a1188
Shorten paper, add research projects
