+++ paper.md
@@ -1,5 +1,5 @@
 ---
-title: '$\Phi_\textrm{ML}$: Intuitive scientific computing with dimension types for Jax, PyTorch, TensorFlow & NumPy'
+title: '$\Phi_\textrm{ML}$: Intuitive Scientific Computing with Dimension Types for Jax, PyTorch, TensorFlow & NumPy'
 tags:
   - Python
   - Machine Learning
@@ -60,183 +60,46 @@ However, current machine learning frameworks have been designed for the core ML 
 This can result in overly verbose code when implementing scientific applications and may require implementing custom operators, since many 
 common functions like sparse-sparse matrix multiplication, periodic padding or sparse linear solvers are not available in all libraries.
 
-$\Phi_\textrm{ML}$ is a scientific computing library based on Python 3 [@Python3] that aims to address these issues and simplify scientific code in the process.
-It consists of a high-level NumPy-like API geared towards writing easy-to-read and scalable simulation code, as well as a neural network API designed to allow users to quickly iterate over many network architectures and hyperparameter settings.
-Similar to eagerpy [@rauber2020eagerpy], $\Phi_\textrm{ML}$ integrates with Jax [@Jax2018], PyTorch [@PyTorch2019], TensorFlow [@TensorFlow2016] and NumPy [@NumPy2020], providing a custom Tensor class.
-However, unlike eagerpy, $\Phi_\textrm{ML}$'s `Tensor` adds additional functionality to make user code more concise and easier to read.
+$\Phi_\textrm{ML}$ is a scientific computing library based on Python 3 [@Python3] targeting scientific applications that use machine learning methods.
+Its main goals are:
 
-$\Phi_\textrm{ML}$ has been in development since 2019 as part of the [$\Phi_\textrm{Flow}$](https://github.com/tum-pbs/PhiFlow) [@phiflow] project where it originated as a unified API for TensorFlow and NumPy, used to run differentiable fluid simulations.
-With $\Phi_\textrm{Flow}$ version 2.0 and consecutive releases, $\Phi_\textrm{ML}$ underwent a drastic overhaul.
-A major issue with the previous API, and in fact all popular ML APIs, is the need for reshaping, which can quickly get out of hand for physical simulations.
-The work towards automatic reshaping sparked most of the changes that have been made to the library since.
+* **Reusability.** Code based on $\Phi_\textrm{ML}$ should be able to run in many settings without modification. It should be agnostic towards the dimensionality of simulated systems and the employed discretization. All code should be trivially vectorizable.
+* **Compatibility.** Users should be free to choose whatever ML or third-party library they desire without modifying their simulation code. $\Phi_\textrm{ML}$ should support Linux, Windows and Mac.
+* **Usability.** $\Phi_\textrm{ML}$ should be easy to learn and use, matching existing APIs where possible. It should encourage users to write concise and expressive code.
+* **Maintainability.** All high-level source code of $\Phi_\textrm{ML}$ should be easy to understand. Continuous testing should be used to ensure that future updates do not break existing code.
+* **Performance.** $\Phi_\textrm{ML}$ should be able to make use of hardware accelerators, such as GPUs and TPUs, where possible. During development, we prioritize rapid code iterations over execution speed but the completed code should run as fast as if written directly against the chosen ML library.
 
-We will first explain the design principles underlying $\Phi_\textrm{ML}$'s development, before detailing the major design decisions and resulting architecture.
-For a list of supported features, see the [GitHub homepage](https://github.com/tum-pbs/PhiML).
+In the following, we explain the architecture and major features that help $\Phi_\textrm{Flow}$ reach these goals.
[…]
