+++ paper.bib
@@ -317,3 +317,136 @@
   year = {2018},
   note = {Accessed on 16 Jan 2024}
 }
+
+
+@article{ScaleInvariant2022,
+  title={Scale-invariant Learning by Physics Inversion},
+  author={Holl, Philipp and Koltun, Vladlen and Thuerey, Nils},
+  journal={Advances in Neural Information Processing Systems},
+  volume={35},
+  pages={5390--5403},
+  year={2022}
+}
+
+@article{HalfInverse2022,
+  title={Half-inverse gradients for physical deep learning},
+  author={Schnell, Patrick and Holl, Philipp and Thuerey, Nils},
+  journal={arXiv preprint arXiv:2203.10131},
+  year={2022}
+}
+
+@article{PBDL2021,
+  title={Physics-based deep learning},
+  author={Thuerey, Nils and Holl, Philipp and Mueller, Maximilian and Schnell, Patrick and Trost, Felix and Um, Kiwon},
+  journal={arXiv preprint arXiv:2109.05237},
+  year={2021}
+}
+
+
+@article{PDEBench,
+  title={PDEBench: An extensive benchmark for scientific machine learning},
+  author={Takamoto, Makoto and Praditia, Timothy and Leiteritz, Raphael and MacKinlay, Daniel and Alesiani, Francesco and Pfl{\"u}ger, Dirk and Niepert, Mathias},
+  journal={Advances in Neural Information Processing Systems},
+  volume={35},
+  pages={1596--1611},
+  year={2022}
+}
+
+@article{PDEArena,
+  title={Towards Multi-spatiotemporal-scale Generalized PDE Modeling},
+  author={Gupta, Jayesh K and Brandstetter, Johannes},
+  journal={arXiv preprint arXiv:2209.15616},
+  year={2022}
+}
+
+
+
+@inproceedings{brandstetter2021message,
+  title={Message Passing Neural PDE Solvers},
+  author={Brandstetter, Johannes and Worrall, Daniel E and Welling, Max},
+  booktitle={International Conference on Learning Representations},
+  year={2021}
+}
+@article{wandel2021teaching,
+  title={Teaching the incompressible Navier--Stokes equations to fast neural surrogate models in three dimensions},
+  author={Wandel, Nils and Weinmann, Michael and Klein, Reinhard},
+  journal={Physics of Fluids},
+  volume={33},
+  number={4},
+  year={2021},
+  publisher={AIP Publishing}
+}
+@inproceedings{brandstetter2022clifford,
+  title={Clifford Neural Layers for PDE Modeling},
+  author={Brandstetter, Johannes and van den Berg, Rianne and Welling, Max and Gupta, Jayesh K},
+  booktitle={The Eleventh International Conference on Learning Representations},
+  year={2022}
+}
+@inproceedings{wandel2020learning,
+  title={Learning Incompressible Fluid Dynamics from Scratch-Towards Fast, Differentiable Fluid Models that Generalize},
+  author={Wandel, Nils and Weinmann, Michael and Klein, Reinhard},
+  booktitle={International Conference on Learning Representations},
+  year={2020}
+}
+@inproceedings{sengar2021multi,
+  title={Multi-task learning based approach for surgical video desmoking},
+  author={Sengar, Vartika and Seemakurthy, Karthik and Gubbi, Jayavardhana and P, Balamuralidhar},
+  booktitle={Proceedings of the twelfth Indian conference on computer vision, graphics and image processing},
+  pages={1--9},
+  year={2021}
+}
+@article{parekh1993sex,
+  title={Sex differences in control of renal outer medullary circulation in rats: role of prostaglandins},
[…]
