### 2024-02-17 #6171 @holl-
Dear @mstimberg, I have modified the paper as you suggested, trimming the previous content to two pages and adding a short discussion about research projects built on this library.

Let me know if there is anything else that needs to be changed!
