### 2024-02-17 #6171 @gauravbokil8
I have been using the u_net function for training some 2D CFD data. It would be quite useful if you could add "strides" and "filter size" as a parameter rather than having them as a default in the function definition. This can add more flexibility to the U-net construction. @holl- @mstimberg Let me know what you think.
