### 2024-02-17 #6171 @holl-
@gauravbokil8 Thanks for pointing out the duplicate cells! Should be fixed now.

I believe exposing `filter_size` should be relatively straightforward. With `stride`, you mean for the down-sampling layers, right? That will complicate the implementations, but I can put it onto the to-do list for the next release. Maybe we could discuss this in a GitHub issue?

When it comes to irregular domains, there are multiple ways this can be implemented. Uniform grid stretching can be done by passing `dx` to spatial functions like `laplace` or `spatial_gradient`. For varying cell stretching, there are correction terms which one can compute. For completely irregular domains, you can use unstructured meshes. One such implementation will be part of Φ-Flow 3.0 and uses sparse matrices to represent the cell connectivity, see [this demo](https://github.com/tum-pbs/PhiFlow/blob/3.0/docs/Unstructured.ipynb). This is work in progress.
