### 2024-02-20 #6171 @holl-
@gauravbokil8 I've added `down_kernel_size` and `up_kernel_size` to [`u_net`](https://tum-pbs.github.io/PhiML/phiml/nn.html#phiml.nn.u_net).

@mstimberg I don't have any more changes planned for the paper. Let me know if anything else needs to be done!
