### 2024-02-20 b4c9a0c
Fix docstring format in nn.py
