+++ phiml/nn.py
@@ -156,11 +156,11 @@ def mlp(in_channels: int,
     Fully-connected neural networks are available in Φ-ML via mlp().
 
     Args:
-        in_channels : size of input layer, int
+        in_channels: size of input layer, int
         out_channels = size of output layer, int
-        layers : tuple of linear layers between input and output neurons, list or tuple
-        activation : activation function used within the layers, string
-        batch_norm : use of batch norm after each linear layer, bool
+        layers: tuple of linear layers between input and output neurons, list or tuple
+        activation: activation function used within the layers, string
+        batch_norm: use of batch norm after each linear layer, bool
 
     Returns:
         Dense net model as specified by input arguments
@@ -183,14 +183,14 @@ def u_net(in_channels: int,
     Built-in U-net architecture, classically popular for Semantic Segmentation in Computer Vision, composed of downsampling and upsampling layers.
 
     Args:
-        in_channels: input channels of the feature map, dtype : int
-        out_channels : output channels of the feature map, dtype : int
-        levels : number of levels of down-sampling and upsampling, dtype : int
-        filters : filter sizes at each down/up sampling convolutional layer, if the input is integer all conv layers have the same filter size,
-        activation : activation function used within the layers, dtype : string
-        batch_norm : use of batchnorm after each conv layer, dtype : bool
-        in_spatial : spatial dimensions of the input feature map, dtype : int
-        use_res_blocks : use convolutional blocks with skip connections instead of regular convolutional blocks, dtype : bool
+        in_channels: input channels of the feature map, dtype: int
+        out_channels: output channels of the feature map, dtype: int
+        levels: number of levels of down-sampling and upsampling, dtype: int
+        filters: filter sizes at each down/up sampling convolutional layer, if the input is integer all conv layers have the same filter size,
+        activation: activation function used within the layers, dtype: string
+        batch_norm: use of batchnorm after each conv layer, dtype: bool
+        in_spatial: spatial dimensions of the input feature map, dtype: int
+        use_res_blocks: use convolutional blocks with skip connections instead of regular convolutional blocks, dtype: bool
         down_kernel_size: Kernel size for convolutions on the down-sampling (first half) side of the U-Net.
         up_kernel_size: Kernel size for convolutions on the up-sampling (second half) of the U-Net.
 
@@ -211,12 +211,12 @@ def conv_net(in_channels: int,
     Built in Conv-Nets are also provided. Contrary to the classical convolutional neural networks, the feature map spatial size remains the same throughout the layers. Each layer of the network is essentially a convolutional block comprising of two conv layers. A filter size of 3 is used in the convolutional layers.
 
     Args:
-        in_channels : input channels of the feature map, dtype : int
-        out_channels : output channels of the feature map, dtype : int
-        layers : list or tuple of output channels for each intermediate layer between the input and final output channels, dtype : list or tuple
-        activation : activation function used within the layers, dtype : string
-        batch_norm : use of batchnorm after each conv layer, dtype : bool
-        in_spatial : spatial dimensions of the input feature map, dtype : int
+        in_channels: input channels of the feature map, dtype: int
+        out_channels: output channels of the feature map, dtype: int
+        layers: list or tuple of output channels for each intermediate layer between the input and final output channels, dtype: list or tuple
+        activation: activation function used within the layers, dtype: string
+        batch_norm: use of batchnorm after each conv layer, dtype: bool
+        in_spatial: spatial dimensions of the input feature map, dtype: int
 
     Returns:
         Conv-net model as specified by input arguments
@@ -237,12 +237,12 @@ def res_net(in_channels: int,
     A default filter size of 3 is used in the convolutional layers.
 
     Args:
[…]
