### 2024-02-25 #6171 @holl-
@mstimberg I have created a release for version 1.4.0 and uploaded the corresponding archive to figshare under the DOI 10.6084/m9.figshare.25282300.
