### 2024-02-27 #6171 @mstimberg
@holl, apologies for the delay, I finally had a close look at the document. From my side it is basically looking good, but there are a few minor issues that I'd like you to fix if possible:
- We don't document this very well, but could you please expand your affiliation into a "Department, University, City, Country" format
- In l. 102 of the current manuscript, could you please use a citation without parentheses (i.e. `@PBDL2021` instead of `[@PBDL2021]`)
- a number of references are missing DOIs, could you please add them where appropriate?
- for arXiv references, could you please use the format given on `arxiv.org` (i.e., use the `@misc` type with `archivePrefix` and `eprint` fields)?

Thanks!
