### 2024-02-27 #6171 @mstimberg
Regarding the figshare archive – we usally ask authors to use the paper title as the title of the archive as well, could you change this accordingly :pray: ? I think on figshare this issues a new DOI (in contrast to zenodo), but this isn't a problem.
