### 2024-02-28 35a457e
Minor paper adjustments
* Affiliation
* PDBL citation
* DOIs and ArXiv format
