+++ paper.bib
@@ -1,16 +1,18 @@
-@article{rauber2020eagerpy,
-  title={{EagerPy}: Writing Code That Works Natively with {PyTorch}, {TensorFlow}, {JAX}, and {NumPy}},
-  author={Rauber, Jonas and Bethge, Matthias and Brendel, Wieland},
-  journal={arXiv preprint arXiv:2008.04175},
+
+@misc{rauber2020eagerpy,
+  title={EagerPy: Writing Code That Works Natively with PyTorch, TensorFlow, JAX, and NumPy},
+  author={Jonas Rauber and Matthias Bethge and Wieland Brendel},
   year={2020},
-  url={https://eagerpy.jonasrauber.de},
+  eprint={2008.04175},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG}
 }
 
-@article{phiflow,
-  title={Learning to control pdes with differentiable physics},
-  author={Holl, Philipp and Koltun, Vladlen and Thuerey, Nils},
-  journal={arXiv preprint arXiv:2001.07457},
-  year={2020}
+@inproceedings{phiflow,
+  title={Learning to Control PDEs with Differentiable Physics},
+  author={Holl, Philipp and Thuerey, Nils and Koltun, Vladlen},
+  booktitle={International Conference on Learning Representations},
+  year={2019}
 }
 
 @book{Python3,
@@ -244,10 +246,11 @@
 
 @inproceedings{TensorFlow2016,
   title={Tensorflow: A system for large-scale machine learning},
-  author={Abadi, Mart{\'\i}n and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
+  author={Abadi, Martin and Barham, Paul and Chen, Jianmin and Chen, Zhifeng and Davis, Andy and Dean, Jeffrey and Devin, Matthieu and Ghemawat, Sanjay and Irving, Geoffrey and Isard, Michael and others},
   booktitle={12th $\{$USENIX$\}$ symposium on operating systems design and implementation ($\{$OSDI$\}$ 16)},
   pages={265--283},
-  year={2016}
+  year={2016},
+  doi={}
 }
 
 @article{PyTorch2019,
@@ -328,18 +331,20 @@
   year={2022}
 }
 
-@article{HalfInverse2022,
-  title={Half-inverse gradients for physical deep learning},
+@inproceedings{HalfInverse2022,
+  title={Half-Inverse Gradients for Physical Deep Learning},
   author={Schnell, Patrick and Holl, Philipp and Thuerey, Nils},
-  journal={arXiv preprint arXiv:2203.10131},
-  year={2022}
+  booktitle={International Conference on Learning Representations},
+  year={2021}
 }
 
-@article{PBDL2021,
-  title={Physics-based deep learning},
-  author={Thuerey, Nils and Holl, Philipp and Mueller, Maximilian and Schnell, Patrick and Trost, Felix and Um, Kiwon},
-  journal={arXiv preprint arXiv:2109.05237},
-  year={2021}
+@misc{PBDL2021,
+  title={Physics-based Deep Learning},
+  author={Nils Thuerey and Philipp Holl and Maximilian Mueller and Patrick Schnell and Felix Trost and Kiwon Um},
+  year={2022},
+  eprint={2109.05237},
+  archivePrefix={arXiv},
+  primaryClass={cs.LG}
 }
 
 
@@ -352,11 +357,13 @@
   year={2022}
 }
 
-@article{PDEArena,
+@misc{PDEArena,
   title={Towards Multi-spatiotemporal-scale Generalized PDE Modeling},
-  author={Gupta, Jayesh K and Brandstetter, Johannes},
[…]
