### 2024-02-28 #6171 @holl-
@mstimberg Thanks for the update!

I've modified the paper as you requested. Not all cited works have associated DOIs, though. I've added all the ones I could find.

The figshare DOI is still `10.6084/m9.figshare.25282300` after the name change.
