### 2024-02-28 #6171 @mstimberg
> I've modified the paper as you requested. Not all cited works have associated DOIs, though. I've added all the ones I could find.

Thanks a lot, I will have a closer look later today (it's perfectly normal that there isn't a DOI for each reference)

> The figshare DOI is still `10.6084/m9.figshare.25282300` after the name change.

It links to https://figshare.com/articles/software/_-ML_1_4_0/25282300 which still seems to state "Φ-ML 1.4.0" as the title, or am I misunderstanding something here? I am not very familiar with figshare myself (all of the articles I edited so far used Zenodo), so it is definitely possible that the mistake is on my end.
