### 2024-02-28 7f5752a
Minor bibliography adjustments/DOI additions
