### 2024-02-28 #6171 @holl-
@mstimberg Thanks for the PR, I merged it.

No, it's not an AAS submission, the metadata was just a leftover from the template. I now removed the entries.

I forgot to publish the changes on figshare before, should be fixed now. The DOI now links to the updated version.
