### 2024-02-29 8b80126
Fix unnecessary slicing for batched solves
