### 2024-02-29 b110400
Check linear trace input
