### 2024-03-01 #6171 @arfon
@holl- the JOSS website (and citation metadata) can't handle TeX. What would a plain-text version of this be? e.g., Phi_ML or something like that: https://joss.theoj.org/papers/10.21105/joss.06171
