### 2024-03-01 #6171 @mstimberg
Ah, that's unfortunate. If Unicode is fine,  Φ_ML would be another option, I guess.
